% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\docType{data}
\name{abnormal_handlers}
\alias{abnormal_handlers}
\alias{IGNORE_HANDLER}
\alias{WARN_HANDLER}
\alias{ERROR_HANDLER}
\title{Handler types for abnormal operations}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
IGNORE_HANDLER

WARN_HANDLER

ERROR_HANDLER
}
\description{
These constants define the different types of handlers for abnormal (but recoverable) operations.
}
\keyword{datasets}
