% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\name{ALL_SCALARS}
\alias{ALL_SCALARS}
\title{Wildcard for all scalars.}
\usage{
ALL_SCALARS
}
\description{
Wildcard for all scalars.
}
\examples{
d <- memory_daf()
set_scalar(d, "organism", "human")
v <- viewer(d, data = list(list(ALL_SCALARS, "=")))
scalars_set(v)
}
