% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\name{ALL_VECTORS}
\alias{ALL_VECTORS}
\title{Wildcard for all vectors.}
\usage{
ALL_VECTORS
}
\description{
Wildcard for all vectors.
}
\examples{
d <- example_cells_daf()
v <- viewer(d, data = list(list(ALL_VECTORS, "=")))
vectors_set(v, "cell")
}
