% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{And}
\alias{And}
\title{And query operation}
\usage{
And(property, ...)
}
\arguments{
\item{property}{String specifying the property}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
A query operation for restricting the set of entries of an \code{Axis}. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.And}{documentation} for details.
}
