% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{axes_set}
\alias{axes_set}
\title{Names of all axes, sorted.}
\usage{
axes_set(daf)
}
\arguments{
\item{daf}{A \code{DafReader}.}
}
\value{
Character vector of axis names.
}
\description{
Names of all axes, sorted.
}
\examples{
# Mirrors readers.jl jldoctest at line 273.
axes_set(example_cells_daf()) # "cell" "donor" "experiment" "gene"
}
