% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Axis}
\alias{Axis}
\title{Axis selection query operation.}
\usage{
Axis(value, ...)
}
\arguments{
\item{...}{Optional piped \link{DafrQuery}.}

\item{axis_name}{Axis name (character scalar), or a piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds an \verb{@ <axis_name>} query fragment, selecting the named axis
as the next input to the query.
}
\examples{
Axis("cell")
Axis("cell") |> Axis("gene")
}
\seealso{
\code{\link[=AsAxis]{AsAxis()}}, \code{\link[=LookupVector]{LookupVector()}}, \code{\link[=LookupMatrix]{LookupMatrix()}}
}
