% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{axis_entries}
\alias{axis_entries}
\title{Entry names of an axis (full or by index).}
\usage{
axis_entries(daf, axis, indices = NULL)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{axis}{Axis name (character scalar).}

\item{indices}{Optional integer index vector (1-based).}
}
\value{
Character vector.
}
\description{
Entry names of an axis (full or by index).
}
