% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{axis_length}
\alias{axis_length}
\title{Length (entry count) of an axis.}
\usage{
axis_length(daf, axis)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{axis}{Axis name (character scalar).}
}
\value{
Integer scalar.
}
\description{
Length (entry count) of an axis.
}
