% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{canonical_query}
\alias{canonical_query}
\title{Canonicalise a query string.}
\usage{
canonical_query(query_string)
}
\arguments{
\item{query_string}{A query string.}
}
\value{
The canonical query string (stable form; suitable for use as cache key).
}
\description{
Canonicalise a query string.
}
\examples{
canonical_query("@ cell : donor")
}
