% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_daf.R
\name{chain_writer}
\alias{chain_writer}
\title{Create a chain of DafReaders with a final DafWriter.}
\usage{
chain_writer(dafs, name = NULL)
}
\arguments{
\item{dafs}{Ordered list of \code{DafReader}s. Later entries override earlier
entries on read.}

\item{name}{Optional chain name; defaults to \code{paste(names, collapse = ";")}.}
}
\value{
A \code{WriteChainDaf}.
}
\description{
Create a chain of DafReaders with a final DafWriter.
}
