% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperCheckDataDirectory.R
\name{CheckDataDirectory}
\alias{CheckDataDirectory}
\title{Checks if the data directory exists in the working directory and throws an error if it doesn't.}
\usage{
CheckDataDirectory()
}
\author{
Ricardo Vladimiro
}
