% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperCheckProjectDirectory.R
\name{CheckProjectDirectory}
\alias{CheckProjectDirectory}
\title{Checks if the project directory exists in the working directory and throws an error if it doesn't.}
\usage{
CheckProjectDirectory()
}
\author{
Ricardo Vladimiro
}

