% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoleClear.R
\name{Clear}
\alias{Clear}
\title{Clear the console because I always forget the proper escape code}
\usage{
Clear()
}
\description{
I'm only writing a description to avoid a warning. I'm a nice package creator. Just type Clear() and the console will be cleared!
}
\author{
Ricardo Vladimiro
}
