% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{Contract}
\alias{Contract}
\title{A contract describing a computation's inputs and outputs.}
\usage{
Contract(name = "", is_relaxed = FALSE, axes = list(), data = list())
}
\arguments{
\item{name}{Optional name.}

\item{is_relaxed}{If TRUE, unknown properties don't error.}

\item{axes}{Named list: axis -> list(expectation, description).}

\item{data}{List of contract_scalar()/contract_vector()/contract_matrix() records.}
}
\description{
A contract describing a computation's inputs and outputs.
}
\examples{
ct <- create_contract(
    scalars = list(contract_scalar("organism", RequiredInput, "character",
                                   "Species name."))
)
inherits(ct, "dafr::Contract")
}
