% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Convert}
\alias{Convert}
\title{Element-wise query operation: convert values to a given type.}
\usage{
Convert(type = NULL, ...)
}
\arguments{
\item{type}{Target type as a character scalar (e.g. \code{"Int64"},
\code{"Float32"}). Positional.}

\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{\% Convert type <value>} query fragment.
}
\examples{
Convert(type = "Int64")
}
\seealso{
\code{\link[=Clamp]{Clamp()}}, \code{\link[=Round]{Round()}}
}
