% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperConvertToCamelCase.R
\name{ConvertToCamelCase}
\alias{ConvertToCamelCase}
\title{Convert a string from under_score syntax to camelCase}
\usage{
ConvertToCamelCase(s)
}
\arguments{
\item{s}{string vector to be converted}
}
\value{
string vector with converted strings
}
\description{
helper function to convert underscore names to camelCase, 
    to ease in following variable naming conventions.
    Copied from https://stackoverflow.com/questions/25504222/elegant-r-function-mixed-case-separated-by-periods-to-underscore-separated-lowe
}
