% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Count}
\alias{Count}
\title{Reduction query operation: count non-zero entries.}
\usage{
Count(type = NULL, ...)
}
\arguments{
\item{type}{Optional output type (character scalar) or a piped
\link{DafrQuery}.}

\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{\% Count} query fragment.
}
\examples{
Count()
}
\seealso{
\code{\link[=Sum]{Sum()}}
}
