% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{CountBy}
\alias{CountBy}
\title{Count-by query operation.}
\usage{
CountBy(value, ...)
}
\arguments{
\item{...}{Optional piped \link{DafrQuery}.}

\item{property}{Property name to count by, or a piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{* <property>} query fragment. Counts entries of the prior
axis, grouped by \code{property}.
}
\examples{
CountBy("donor")
}
\seealso{
\code{\link[=GroupBy]{GroupBy()}}
}
