% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectCreateProjectStructure.R
\name{CreateProjectStructure}
\alias{CreateProjectStructure}
\title{Create base directories and files.}
\usage{
CreateProjectStructure()
}
\description{
Creates directory code and data and git ignore file. The git ignore file is from https://github.com/github/gitignore/blob/master/R.gitignore plus any additions from us.
}
\author{
Ricardo Vladimiro
}

