% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daf.R
\name{Daf}
\alias{Daf}
\title{Daf (Data Axes Format) S3 object}
\usage{
Daf(jl_obj)
}
\arguments{
\item{jl_obj}{The Julia DafReader object to wrap}
}
\value{
A Daf S3 object
}
\description{
An S3 object that wraps a Julia DataAxesFormats object and provides access to its methods.
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/data.html#DataAxesFormats.Data.DafReader}{documentation} for details.
}
