% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{daf_name}
\alias{daf_name}
\title{Return the name of a \code{DafReader}.}
\usage{
daf_name(x)
}
\arguments{
\item{x}{A \link{DafReader}.}
}
\value{
Character scalar.
}
\description{
Asserts that \code{x} is a \link{DafReader} and returns its \code{name} property
(the string passed to the constructor).
}
\examples{
daf_name(memory_daf(name = "hello"))
}
\seealso{
\code{\link[=is_daf]{is_daf()}}, \code{\link[=read_only]{read_only()}}
}
