% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dafr-package.R
\docType{package}
\name{dafr-package}
\alias{dafr}
\alias{dafr-package}
\title{dafr: Multi-Dimensional Data Along Arbitrary Axes}
\description{
A native R implementation of the DataAxesFormats (DAF) data model for multi-dimensional data along arbitrary axes. Provides the MemoryDaf and FilesDaf storage backends, a query DSL, views, chains, contracts, computations, adapters, and operations. Designed for efficiency on large single-cell datasets via memory-mapped file access (ALTREP), hand-tuned C++ kernels, and optional OpenMP parallelism.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tanaylab/dafr}
  \item Report bugs at \url{https://github.com/tanaylab/dafr/issues}
}

}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il} (\href{https://orcid.org/0000-0002-8458-9507}{ORCID})

Authors:
\itemize{
  \item Oren Ben-Kiki \email{oren@ben-kiki.org}
}

Other contributors:
\itemize{
  \item Weizmann Institute of Science [copyright holder]
}

}
\keyword{internal}
