% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{DafReadOnly}
\alias{DafReadOnly}
\title{Abstract read-only reader class.}
\usage{
DafReadOnly(
  name = character(0),
  internal = new.env(parent = emptyenv()),
  cache = new.env(parent = emptyenv()),
  axis_version_counter = new.env(parent = emptyenv()),
  vector_version_counter = new.env(parent = emptyenv()),
  matrix_version_counter = new.env(parent = emptyenv())
)
}
\description{
Abstract read-only reader class.
}
