% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{delete_axis}
\alias{delete_axis}
\title{Delete an axis (and all vectors / matrices that depend on it).}
\usage{
delete_axis(daf, axis, must_exist = TRUE)
}
\arguments{
\item{daf}{A \code{DafWriter}.}

\item{axis}{Axis name.}

\item{must_exist}{If \code{TRUE} (default) raise when the axis is absent;
if \code{FALSE} silently no-op.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
Delete an axis (and all vectors / matrices that depend on it).
}
\examples{
d <- memory_daf()
add_axis(d, "cell", c("c1", "c2"))
add_axis(d, "gene", c("g1", "g2", "g3"))
delete_axis(d, "gene")
axes_set(d)
}
