% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{delete_matrix}
\alias{delete_matrix}
\title{Delete a matrix.}
\usage{
delete_matrix(daf, rows_axis, columns_axis, name, must_exist = TRUE)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{rows_axis}{Row-axis name.}

\item{columns_axis}{Column-axis name.}

\item{name}{Matrix name.}

\item{must_exist}{See \code{delete_axis}.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
Delete a matrix.
}
\examples{
d <- memory_daf()
add_axis(d, "cell", c("c1", "c2"))
add_axis(d, "gene", c("g1", "g2", "g3"))
m <- matrix(0L, nrow = 2, ncol = 3)
set_matrix(d, "cell", "gene", "counts", m)
delete_matrix(d, "cell", "gene", "counts")
has_matrix(d, "cell", "gene", "counts")
}
