% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{delete_scalar}
\alias{delete_scalar}
\title{Delete a scalar.}
\usage{
delete_scalar(daf, name, must_exist = TRUE)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{name}{Scalar name.}

\item{must_exist}{See \code{delete_axis}.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
Delete a scalar.
}
