% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{delete_vector}
\alias{delete_vector}
\title{Delete a vector on an axis.}
\usage{
delete_vector(daf, axis, name, must_exist = TRUE)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{axis}{Axis name.}

\item{name}{Vector name.}

\item{must_exist}{See \code{delete_axis}.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
Delete a vector on an axis.
}
\examples{
d <- memory_daf()
add_axis(d, "cell", c("c1", "c2"))
set_vector(d, "cell", "donor", c("d1", "d2"))
delete_vector(d, "cell", "donor")
has_vector(d, "cell", "donor")
}
