% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{empty_cache}
\alias{empty_cache}
\title{Empty caches on a Daf object.}
\usage{
empty_cache(daf, group = c("mapped", "memory", "query"))
}
\arguments{
\item{daf}{A \code{DafReader}/\code{DafWriter} instance.}

\item{group}{Character vector; any of \code{"mapped"}, \code{"memory"}, \code{"query"}.
Defaults to all three.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
Empty caches on a Daf object.
}
