% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{EndMask}
\alias{EndMask}
\title{End-mask query operation.}
\usage{
EndMask(...)
}
\arguments{
\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{]} query fragment that closes a masked subquery opened
with \code{\link[=BeginMask]{BeginMask()}} or \code{\link[=BeginNegatedMask]{BeginNegatedMask()}}.
}
\examples{
EndMask()
Axis("cell") |> BeginMask("type") |> EndMask()
}
\seealso{
\code{\link[=BeginMask]{BeginMask()}}, \code{\link[=BeginNegatedMask]{BeginNegatedMask()}}
}
