% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_chain_daf}
\alias{example_chain_daf}
\title{Load a chain of cells and metacells example data.}
\usage{
example_chain_daf(name = "chain!")
}
\arguments{
\item{name}{Name for the returned \code{WriteChainDaf}.}
}
\value{
A \code{\link{WriteChainDaf}} combining \code{example_cells_daf()} and
\code{example_metacells_daf()}.
}
\description{
Replicates the Julia \code{ExampleData.example_chain_daf()} dataset.
}
\examples{
ch <- example_chain_daf()
axes_set(ch)
}
