% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{expectation-constants}
\alias{expectation-constants}
\alias{RequiredInput}
\alias{OptionalInput}
\alias{CreatedOutput}
\alias{GuaranteedOutput}
\alias{OptionalOutput}
\title{Contract expectation constants}
\usage{
RequiredInput

OptionalInput

CreatedOutput

GuaranteedOutput

OptionalOutput
}
\value{
Character scalar.
}
\description{
String literals used in \code{contract_scalar()} /
\code{contract_vector()} / \code{contract_matrix()} + axis specs.
}
