% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_empty_sparse_matrix}
\alias{get_empty_sparse_matrix}
\title{Get an empty sparse matrix for filling}
\usage{
get_empty_sparse_matrix(
  daf,
  rows_axis,
  columns_axis,
  name,
  eltype,
  nnz,
  indtype = "Int64",
  overwrite = FALSE
)
}
\arguments{
\item{daf}{A Daf object}

\item{rows_axis}{Name of the rows axis}

\item{columns_axis}{Name of the columns axis}

\item{name}{Name of the matrix property}

\item{eltype}{Element type for the matrix values (e.g., "Float64", "Int32")}

\item{nnz}{Number of non-zero elements expected}

\item{indtype}{Optional index type (e.g., "Int32"). If NULL, the default is used.}

\item{overwrite}{Whether to overwrite if matrix already exists (FALSE by default)}
}
\value{
A Julia sparse matrix object that can be filled in-place
}
\description{
Returns an empty sparse matrix for the specified axes and property, which can be filled
in-place. This is useful for efficiently constructing large sparse matrices.
After filling with \code{filled_empty_sparse_matrix}, the matrix is stored in the Daf object.
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/writers.html#DataAxesFormats.Writers.get_empty_sparse_matrix!}{documentation} for details.
}
