% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_empty_sparse_vector}
\alias{get_empty_sparse_vector}
\title{Get an empty sparse vector for filling}
\usage{
get_empty_sparse_vector(
  daf,
  axis,
  name,
  eltype,
  nnz,
  indtype = "Int64",
  overwrite = FALSE
)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis}

\item{name}{Name of the vector property}

\item{eltype}{Element type for the vector values (e.g., "Float64", "Int32")}

\item{nnz}{Number of non-zero elements expected}

\item{indtype}{Optional index type (e.g., "Int32"). If NULL, the default is used.}

\item{overwrite}{Whether to overwrite if vector already exists (FALSE by default)}
}
\value{
A Julia sparse vector object that can be filled in-place
}
\description{
Returns an empty sparse vector for the specified axis and property, which can be filled
in-place. This is useful for efficiently constructing large sparse vectors.
After filling with \code{filled_empty_sparse_vector}, the vector is stored in the Daf object.
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/writers.html#DataAxesFormats.Writers.get_empty_sparse_vector!}{documentation} for details.
}
