% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{get_vector}
\alias{get_vector}
\title{Get a vector, returning it as an axis-named R vector.}
\usage{
get_vector(daf, axis, name, default)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{axis}{Axis name.}

\item{name}{Vector name.}

\item{default}{If supplied and the vector is absent, return a
constant-valued named vector of length \code{axis_length(daf, axis)}
with the axis entries as names.}
}
\value{
Named atomic vector.
}
\description{
Get a vector, returning it as an axis-named R vector.
}
