% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testBootstrapConfidenceInterval.R
\name{GetBootstrappedCI}
\alias{GetBootstrappedCI}
\title{Constructs Confidence Interval for difference in means}
\usage{
GetBootstrappedCI(dataset, testGroups, controlGroup = "Control",
  responseColumn = "response", groupColumn = "abGroup",
  bootIterations = 999, bootConfLevel = 95)
}
\arguments{
\item{dataset}{The dataset containing one observation per row}

\item{testGroups}{A character vector containing the names of the treatement groups}

\item{controlGroup}{The name which identifies the control group in the dataset}

\item{responseColumn}{The name of the column that contains the measure to be tested}

\item{groupColumn}{The name of the column that contains the group identification}

\item{bootIterations}{The number of bootstrap resamples to perform}

\item{bootConfLevel}{The confidence level for the interval. Between 1 and 100}
}
\value{
List containing two dataframes. The first dataframe contains the boundaries of the CI. The second dataframe contains all calculated estimates for the mean of each group
}
\description{
Receives a dataset with a response column, containing the measure for which the confidence interval will be constructed.
The confidence interval will always refer to the difference in means between a treatment gorup and the control group.
It is expected for the dataset to contain one measure and group assignment per observation

This function should also receive a character vector that identifies the name of the control group, as well as a vector of strings containing the names of the treatment groups to compare against the control
The Name of the column containing the measure for the confidence interval should also be provided

Each bootstrap is performed via the smean.cl.boot on the Hmisc package.

Returns the calculated statistics for each bootstrap iteration as well as the boundaries for the confidence interval
}
\author{
João Monteiro
}

