% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperGetDBConfig.R
\name{GetDBConfig}
\alias{GetDBConfig}
\title{Get database server configuration from a given a yaml configuration file}
\usage{
GetDBConfig(id, yamlFile)
}
\arguments{
\item{id}{The id in the yaml config file}

\item{yamlFile}{The configuration yaml file}
}
\value{
A list of configuration parameters
}
\description{
Get database server configuration from a given a yaml configuration file
}
\author{
Ricardo Vladimiro
}
