% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperGetQueryStatement.R
\name{GetQueryStatement}
\alias{GetQueryStatement}
\title{Get a clean query statement from a SQL file or string}
\usage{
GetQueryStatement(query)
}
\arguments{
\item{query}{The query to be ran. It can be a file with SQL or a character vector}
}
\value{
A character vector ready to be used on queries
}
\author{
Ricardo Vladimiro
}

