% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperGetS3Config.R
\name{GetS3Config}
\alias{GetS3Config}
\title{Get S3 server configuration from a given a yaml configuration file}
\usage{
GetS3Config(id, yamlFile)
}
\arguments{
\item{id}{The server name in the yaml config file}

\item{yamlFile}{The yaml configuration file with the S3 information}
}
\value{
A list of configuration parameters
}
\description{
Get S3 server configuration from a given a yaml configuration file
}
\author{
Ricardo Vladimiro
}
