% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperGetYAMLConfig.R
\name{GetYAMLConfig}
\alias{GetYAMLConfig}
\title{Get a configuration from a yaml configuration file}
\usage{
GetYAMLConfig(id, yamlFile)
}
\arguments{
\item{id}{The id in the yaml config file}

\item{yamlFile}{The yaml configuration file with the configuration}
}
\value{
A list of configuration parameters
}
\author{
Ricardo Vladimiro
}

