% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{GroupRowsBy}
\alias{GroupRowsBy}
\title{Group-rows-by query operation.}
\usage{
GroupRowsBy(value, ...)
}
\arguments{
\item{...}{Optional piped \link{DafrQuery}.}

\item{property}{Property name to group rows by, or a piped
\link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{-/ <property>} query fragment. Groups rows of a matrix by
\code{property}; typically followed by \code{\link[=ReduceToRow]{ReduceToRow()}}.
}
\examples{
GroupRowsBy("donor")
}
\seealso{
\code{\link[=GroupBy]{GroupBy()}}, \code{\link[=GroupColumnsBy]{GroupColumnsBy()}}
}
