% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\docType{data}
\name{handler-constants}
\alias{handler-constants}
\alias{ERROR_HANDLER}
\alias{WARN_HANDLER}
\alias{IGNORE_HANDLER}
\title{Inefficient-action handler constants.}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
ERROR_HANDLER

WARN_HANDLER

IGNORE_HANDLER
}
\value{
Character scalar (one of \code{"error"}, \code{"warn"}, \code{"ignore"}).
}
\description{
String constants matching the lowercase tokens accepted by
\code{\link[=register_dafr_handler]{register_dafr_handler()}}. Pass these to
\code{\link[=inefficient_action_handler]{inefficient_action_handler()}} or any registry entry that takes
an action token.
}
\examples{
inefficient_action_handler(WARN_HANDLER)
}
\keyword{datasets}
