% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{has_axis}
\alias{has_axis}
\title{Test whether an axis exists.}
\usage{
has_axis(daf, axis)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{axis}{Axis name (character scalar).}
}
\value{
Logical scalar.
}
\description{
Test whether an axis exists.
}
\examples{
# Mirrors readers.jl jldoctest at lines 210 + 218.
has_axis(example_cells_daf(),     "metacell") # FALSE
has_axis(example_metacells_daf(), "metacell") # TRUE
}
