% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{has_scalar}
\alias{has_scalar}
\title{Test whether a scalar exists.}
\usage{
has_scalar(daf, name)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{name}{Scalar name.}
}
\value{
Logical scalar.
}
\description{
Test whether a scalar exists.
}
