% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{has_vector}
\alias{has_vector}
\title{Test whether a vector exists on an axis.}
\usage{
has_vector(daf, axis, name)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{axis}{Axis name.}

\item{name}{Vector name.}
}
\value{
Logical scalar.
}
\description{
Test whether a vector exists on an axis.
}
\examples{
d <- memory_daf()
add_axis(d, "cell", c("c1", "c2"))
set_vector(d, "cell", "donor", c("d1", "d2"))
has_vector(d, "cell", "donor")
has_vector(d, "cell", "age")
}
