% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{inefficient_action_handler}
\alias{inefficient_action_handler}
\title{Register a handler for the \code{"inefficient"} action category.}
\usage{
inefficient_action_handler(handler)
}
\arguments{
\item{handler}{One of \code{\link[=handler-constants]{ERROR_HANDLER}},
\code{\link[=handler-constants]{WARN_HANDLER}},
\code{\link[=handler-constants]{IGNORE_HANDLER}}, or a function
\verb{function(message, ...)}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Thin wrapper around \code{register_dafr_handler("inefficient", handler)}.
Exists to match the Julia-facade wrapper's API.
}
\examples{
inefficient_action_handler(IGNORE_HANDLER)
inefficient_action_handler(function(msg) message("inefficient: ", msg))
}
