% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquireInsertRedshiftTableByQuery.R
\name{InsertRedshiftTableByQuery}
\alias{InsertRedshiftTableByQuery}
\title{Creates a table with the results of a given SQL query}
\usage{
InsertRedshiftTableByQuery(query, tableName, schema = "development",
  clean = TRUE, id, yamlFile = "../db.yml")
}
\arguments{
\item{query}{Character vector with length 1. Can be either a SQL query or a path to a text file containing a SQL query}

\item{tableName}{The name of the table to create or append to}

\item{schema}{Character vector with length 1 with the name of the DB schema were the table will be created}

\item{clean}{Boolean indicating whether or not to remove existing records from the table before inserting}

\item{id}{The name of the yaml group containing the database credentials}

\item{yamlFile}{The path to the yaml file}
}
\value{
The function will return a boolean value indicating if the process completed successfully
}
\description{
Given a SQL query, this functions inserts it's results onto a redshift table.
We can choose to clean the existing table records before inserting or to simply append.
If the table does not exist, one will be automatically created

The yaml file must contain host, dbname, user and password.
}
\author{
João Monteiro
}

