% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{is_axis_query}
\alias{is_axis_query}
\title{Test whether a query yields an axis entry vector.}
\usage{
is_axis_query(query_string)
}
\arguments{
\item{query_string}{A query string.}
}
\value{
Logical scalar.
}
\description{
Test whether a query yields an axis entry vector.
}
\examples{
is_axis_query("@ cell")
is_axis_query(". organism")
}
