% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_daf}
\alias{is_daf}
\title{Test whether an object is a \code{DafReader}.}
\usage{
is_daf(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
\code{TRUE} if \code{x} inherits from \link{DafReader}, else \code{FALSE}.
}
\description{
Non-throwing predicate for any of the S7 class descendants
(\code{MemoryDaf}, \code{FilesDaf}, \code{ReadOnlyChainDaf}, \code{WriteChainDaf},
\code{ViewDaf}, \code{ContractDaf}, ...).
}
\examples{
is_daf(memory_daf())
is_daf(NULL)
}
