% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{IsNotMatch}
\alias{IsNotMatch}
\title{Negated regex-match query operation.}
\usage{
IsNotMatch(value, ...)
}
\arguments{
\item{value}{Regex pattern (character scalar), or a piped \link{DafrQuery}.}

\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{!~ <pattern>} query fragment. Filters entries whose property
value does NOT match the given regular-expression pattern.
}
\examples{
IsNotMatch("^unknown")
Axis("cell") |> BeginMask("type") |> IsNotMatch("^unknown") |> EndMask()
}
\seealso{
\code{\link[=IsMatch]{IsMatch()}}, \code{\link[=IsNotEqual]{IsNotEqual()}}
}
