% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Lookup}
\alias{Lookup}
\title{Lookup query operation}
\usage{
Lookup(property, ...)
}
\arguments{
\item{property}{String specifying the property name}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
A query operation for looking up the value of a property with some name.
See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.Lookup}{documentation}
for details.
}
