% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{MaskSlice}
\alias{MaskSlice}
\title{MaskSlice query operation}
\usage{
MaskSlice(axis, ...)
}
\arguments{
\item{axis}{String specifying the axis}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
A query operation for using a slice of a matrix as a mask, when the other axis of the matrix is different from the
mask axis. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.MaskSlice}{documentation} for
details.
}
