% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{matrix_contract}
\alias{matrix_contract}
\title{Create a matrix contract specification}
\usage{
matrix_contract(rows_axis, cols_axis, name, expectation, dtype, description)
}
\arguments{
\item{rows_axis}{Rows axis name}

\item{cols_axis}{Columns axis name}

\item{name}{Matrix name}

\item{expectation}{One of RequiredInput, OptionalInput, GuaranteedOutput, OptionalOutput}

\item{dtype}{Data type (e.g., "numeric", "integer")}

\item{description}{Human-readable description}
}
\value{
Matrix specification list
}
\description{
Create a matrix contract specification
}
