% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Max}
\alias{Max}
\title{Reduction query operation: maximum.}
\usage{
Max(...)
}
\arguments{
\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{\% Max} query fragment. Consumes a matrix axis (or the
single axis of a vector) and produces a reduced result.
}
\examples{
Max()
}
\seealso{
\code{\link[=Min]{Min()}}, \code{\link[=Mean]{Mean()}}, \code{\link[=Sum]{Sum()}}
}
