% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Mean}
\alias{Mean}
\title{Reduction query operation: mean.}
\usage{
Mean(type = NULL, ...)
}
\arguments{
\item{type}{Optional output type (character scalar) or a piped
\link{DafrQuery}.}

\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{\% Mean} query fragment (optionally \verb{\% Mean type: <T>}).
}
\examples{
Mean()
Mean(type = "Float64")
}
\seealso{
\code{\link[=Sum]{Sum()}}, \code{\link[=Median]{Median()}}
}
