% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Median}
\alias{Median}
\title{Reduction query operation: median.}
\usage{
Median(type = NULL, ...)
}
\arguments{
\item{type}{Optional output type (character scalar) or a piped
\link{DafrQuery}.}

\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{\% Median} query fragment.
}
\examples{
Median()
}
\seealso{
\code{\link[=Mean]{Mean()}}, \code{\link[=Quantile]{Quantile()}}
}
