% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\docType{data}
\name{merge_actions}
\alias{merge_actions}
\alias{MERGE_SKIP}
\alias{MERGE_LAST_VALUE}
\alias{MERGE_COLLECT_AXIS}
\title{Merge action constants for \code{\link[=concatenate]{concatenate()}}.}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
MERGE_SKIP

MERGE_LAST_VALUE

MERGE_COLLECT_AXIS
}
\description{
Used as values in \code{concatenate()}'s \code{merge} argument:
\code{MERGE_SKIP} skips the property (default), \code{MERGE_LAST_VALUE} uses the
last source containing it, \code{MERGE_COLLECT_AXIS} collects values along
the \code{dataset_axis}.
}
\keyword{datasets}
