% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Min}
\alias{Min}
\title{Reduction query operation: minimum.}
\usage{
Min(...)
}
\arguments{
\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{\% Min} query fragment.
}
\examples{
Min()
}
\seealso{
\code{\link[=Max]{Max()}}, \code{\link[=Mean]{Mean()}}, \code{\link[=Sum]{Sum()}}
}
