% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Mode}
\alias{Mode}
\title{Reduction query operation: most-frequent value.}
\usage{
Mode(type = NULL, ...)
}
\arguments{
\item{type}{Optional output type (character scalar) or a piped
\link{DafrQuery}.}

\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{\% Mode} query fragment. Returns the most common value of
the reduced axis. Numeric-only.
}
\examples{
Mode()
}
\seealso{
\code{\link[=Median]{Median()}}
}
