% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{name}
\alias{name}
\title{Gets the name of a Daf object}
\usage{
name(x, ...)
}
\arguments{
\item{x}{A Daf object}

\item{...}{Additional arguments (not used)}
}
\value{
The name of the Daf data set as a character string
}
\description{
Returns the unique identifier name of the Daf data set.
}
\details{
Each Daf data set has a unique name used in error messages and for identification.
This is typically set when creating the object or derived from its contents.
}
