% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Names}
\alias{Names}
\title{Names query operation.}
\usage{
Names(...)
}
\arguments{
\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{?} query fragment, listing the entries of the prior axis.
}
\examples{
Names()
Axis("cell") |> Names()
}
\seealso{
\code{\link[=Axis]{Axis()}}
}
