% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Or}
\alias{Or}
\title{Or query operation}
\usage{
Or(property, ...)
}
\arguments{
\item{property}{String specifying the property to use for expanding the selection}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
A query operation for expanding the set of entries of an axis.
This operation adds entries to the current selection where the specified property
contains true values (or non-zero/non-empty values). It performs a logical OR
between the current selection and the specified property, treating the property
as a Boolean mask.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.Or}{documentation} for details.
}
