% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{OrNot}
\alias{OrNot}
\title{OrNot query operation}
\usage{
OrNot(property, ...)
}
\arguments{
\item{property}{String specifying the property}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Same as \code{Or} but use the inverse of the mask. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.OrNot}{documentation} for details.
}
