% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperParallelMeanBoot.R
\name{parallelMeanBoot}
\alias{parallelMeanBoot}
\title{Get Mean of vector by bootstrap}
\arguments{
\item{x}{numerical vector}

\item{conf.int}{confidence interval for the estimation of the mean}

\item{B}{number of resamplings for the bootstrapp process}

\item{na.rm}{removes NAs from x. Defaults to TRUE}

\item{reps}{if TRUE the bootstrapped means will be returned as the reps attribute of the returned object}
}
\value{
vector of the mean, and upper and lower estimates
}
\description{
Parallelized implementation of the smean.cl.boot function from the Hmisc package
}
\author{
João Monteiro
}
