% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_parse.R
\name{parse_query}
\alias{parse_query}
\title{Parse a query string into an AST (list of \code{qop} nodes).}
\usage{
parse_query(query_string)
}
\arguments{
\item{query_string}{A character scalar.}
}
\value{
A list of AST node records.
}
\description{
Parse a query string into an AST (list of \code{qop} nodes).
}
\examples{
# Most users call get_query() directly; parse_query() returns the AST.
ast <- parse_query("@ cell : donor")
is_axis_query("@ cell : donor")
get_query(example_cells_daf(), "@ cell : donor") |> head()
}
