% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCohortRetentionRates.R
\name{PerformCohortRetentionTests}
\alias{PerformCohortRetentionTests}
\title{Compares treatment to control on retention rate for given retention day}
\usage{
PerformCohortRetentionTests(dataset, retentionDays, controlName = "Control",
  groupColumn = "abGroup", retentionDaysColumn = "retentionDay",
  usersColumn = "users", acquisitionColumn = "cohort", ...)
}
\arguments{
\item{dataset}{The summarised dataset}

\item{retentionDays}{An integer vector containing the retention days to test. If not provided will test all avaliable non-zero retention days}

\item{controlName}{The name which identifies the control group in the dataset}

\item{groupColumn}{The name of the column that contains the group identification}

\item{retentionDaysColumn}{The name of the column that contains the retention day identification}

\item{usersColumn}{The name of the column that contains the sum of users for each combination of acquisition and activity dates}

\item{acquisitionColumn}{The name of the column that contains the dates of acquisition}

\item{...}{Additional arguments to pass on to prop.test function}
}
\value{
A dataframe containing p-value for two-tailed test, boundaries for confidence intervals, for each tested retentention day and treatment group
}
\description{
This function receives a summarised dataset and compares the retention rates for each avaliable test group against the control.
The dataset should have the columns containing the following information:
acquisition_date | activity_date | retention_day | group | active_users
The names of the columns can differ

This function should also receive a character vector that identifies the name of the control group, as well as the names of the columns containing metrics of interest

This function uses the prop.test function to perform a simple chi^2 on the proportion of users who return N days after their acquisition.
If no retention days are provided, then the function will perform tests for each one avaliable.

Returns p-value for two-tailed test and statistics for each comparison to control
}
\author{
João Monteiro
}
