% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDailyMonetisationKpis.R
\name{PerformDailyMonetisationTests}
\alias{PerformDailyMonetisationTests}
\title{Compares treatment to control on ARPDAU, ARPPU and Conversion Rate}
\usage{
PerformDailyMonetisationTests(dataset, kpi, revenueColumn = "revenue",
  groupColumn = "abGroup", controlName = "Control", bootIterations = 999,
  bootConfLevel = 95, ...)
}
\arguments{
\item{dataset}{The dataset containing one observation per row}

\item{kpi}{One of the following: 'arpdau', 'arppu', 'conversion'}

\item{revenueColumn}{The name of the column that contains the measures on revenue per observation}

\item{groupColumn}{The name of the column that contains the group identification}

\item{controlName}{The name which identifies the control group in the dataset}

\item{bootIterations}{The number of bootstrap resamples to perform}

\item{bootConfLevel}{The confidence level for the interval. Between 1 and 100}

\item{...}{Additional arguments to pass on to permTS function}
}
\value{
List containing two dataframes. The first dataset contains all the p-values and CI boundaries. The second dataframe contains all calculated estimates for the mean of each group
}
\description{
Receives a dataset with a response column. It is expected for this dataset to contain one row per DAU, including non-paying users.
The function provides one of three options for measures to test and construct confidence intervals:
arpdau: this will use the entirety of the dataset
arppu: this will subset the datset to users whose revenue > 0
conversion: this will create a biinary column, on which 1 identifies a user who did a purchase

This function should also receive a character vector that identifies the name of the control group, as well as the name of the column that contains the revenue for each DAU

This functions combines the functions PerformPermutation and GetBootstrappedCI to test the most common monetisation KPIs against all avaliable treatement groups

Returns all the p-values and confidence interval boundaries for each treatement. It also returns the calculated statistics for each bootstrap iteration
}
\author{
João Monteiro
}

