% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPermutations.R
\name{PerformPermutationTest}
\alias{PerformPermutationTest}
\title{Performs a permutation test}
\usage{
PerformPermutationTest(dataset, testGroups, controlGroup = "Control",
  responseColumn = "response", groupColumn = "abGroup", ...)
}
\arguments{
\item{dataset}{The dataset containing one observation per row}

\item{testGroups}{A character vector containing the names of the treatement groups}

\item{controlGroup}{The name which identifies the control group in the dataset}

\item{responseColumn}{The name of the column that contains the measure to be tested}

\item{groupColumn}{The name of the column that contains the group identification}

\item{...}{Additional arguments to pass on to permTS function}
}
\value{
A dataframe containing p-values and statistics for each comparion to control
}
\description{
Receives a dataset with a response column, containing the measure to be tested
It is expected for the dataset to contain one measure and group assignment per observation

This function should also receive a character vector that identifies the name of the control group, as well as a vector of strings containing the names of the treatment groups to compare against the control
The Name of the column containing the measure to be tested should also be provided

Each individual permutation test is performed via the permTS function, from the perm package

Returns all p-values and statistics for each comparison to control
}
\author{
João Monteiro
}

