% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daf.R
\name{print.Daf}
\alias{print.Daf}
\title{Print method for Daf objects}
\usage{
\method{print}{Daf}(x, ...)
}
\arguments{
\item{x}{The Daf object to print}

\item{...}{Additional arguments passed to print}
}
\value{
The Daf object (invisibly)
}
\description{
Prints a description of the Daf object using Julia's description function
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/data.html#DataAxesFormats.Data.description}{documentation} for details.
}
