% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{print.DafContract}
\alias{print.DafContract}
\title{Print method for DafContract}
\usage{
\method{print}{DafContract}(x, ...)
}
\arguments{
\item{x}{A DafContract object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for DafContract
}
