% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquireRedshiftQuery.R
\name{RedshiftQuery}
\alias{RedshiftQuery}
\title{Query data from a Redshift dbID}
\usage{
RedshiftQuery(query, dbID, s3ID, yamlConfig = "../db.yml")
}
\description{
Describe this thing
Get the clean query statement
- The 1st GetQueryStatement will return the intended query
- Next we escape quotes
- The 2nd GetQueryStatement will return the version with the unload code.
}
\author{
Henrique Cabral and Ricardo Vladimiro
}

