% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{registered_eltwise}
\alias{registered_eltwise}
\title{List registered eltwise operations.}
\usage{
registered_eltwise()
}
\value{
Sorted character vector of registered eltwise names.
}
\description{
Returns the names of all currently registered eltwise operations.
}
