% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperReplaceQueryParams.R
\name{ReplaceQueryParams}
\alias{ReplaceQueryParams}
\title{Replaces \%s parameters in a query statement by the given arguments}
\arguments{
\item{query}{The query to be treated It can be a file with SQL or a character vector}

\item{...}{a number of strings to be replaced in the query}
}
\value{
A character vector ready to be used on queries
}
\description{
Replaces \%s parameters in a query statement by the given arguments
}
\examples{
ReplaceQueryParams(
    query = "SELECT * FROM \%s_session",
    "game_name"
)

ReplaceQueryParams(
    query = "SELECT *, '\%1$s' AS game 
    FROM \%1$s_session
    WHERE date = '\%2$s'",
    "game_name",
    Sys.Date()
)
}
\author{
João Monteiro
}
