% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\docType{data}
\name{RequiredInput}
\alias{RequiredInput}
\alias{OptionalInput}
\alias{GuaranteedOutput}
\alias{OptionalOutput}
\title{Contract Expectation Types}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
RequiredInput

OptionalInput

GuaranteedOutput

OptionalOutput
}
\description{
These constants correspond to Julia's ContractExpectation enum
}
\keyword{datasets}
