% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Round}
\alias{Round}
\title{Element-wise query operation: round to nearest integer.}
\usage{
Round(...)
}
\arguments{
\item{...}{Optional \link{DafrQuery} pipe target.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{\% Round} query fragment.
}
\examples{
Round()
}
\seealso{
\code{\link[=Abs]{Abs()}}, \code{\link[=Significant]{Significant()}}
}
