% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquireS3GetUnload.R
\name{S3GetUnload}
\alias{S3GetUnload}
\title{Read s3 files generated from Unload SQL command}
\usage{
S3GetUnload(filePrefix, dbID, s3ID, yamlConfig = "../db.yml",
  acceleration = TRUE, parallel = TRUE)
}
\arguments{
\item{filePrefix}{Either a character vector with length 1, with the prefix shared by the objects to be read, or a character vector listing the full path of each obejct to be read}

\item{dbID}{The name of the yaml group containing the database credentials}

\item{s3ID}{The name of the yaml group containing the s3 bucket credentials}

\item{yamlConfig}{The path to the yaml file}

\item{acceleration}{Use s3 bucket acceleration}

\item{parallel}{Use parallel package for simultaneous connections to s3}
}
\value{
The function will return a data.table object with the content of the files that were read from the s3 bucket
}
\description{
This function reads into a data table the files generated from an unload command.
The files that are read in, are the ones that share the prefix passed on the argument
The prefix should contain the absolute path to the s3 folder holding the files
alternatively we can pass the function a character vector listing all the files to be read

The yaml file must contain s3 bucket name, user and password.
}
\author{
João Monteiro
}

