% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSaveProjectEnvironment.R
\name{SaveProjectEnvironment}
\alias{SaveProjectEnvironment}
\title{Saves the project environment in the project directory}
\usage{
SaveProjectEnvironment()
}
\description{
Saves the project environment in the project directory
}
\author{
Ricardo Vladimiro
}
