% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoleSay.R
\name{Say}
\alias{Say}
\title{Say something in the console}
\usage{
Say(...)
}
\arguments{
\item{...}{Variables to be sent to the console}
}
\description{
Say has the same behavior as cat but adds a carriage return to the end. It will output a sentence built from the arguments passed. A carriage return is added for ease of use.
}
\author{
Ricardo Vladimiro
}
