% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{set_scalar}
\alias{set_scalar}
\title{Set a scalar.}
\usage{
set_scalar(daf, name, value, overwrite = FALSE)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{name}{Scalar name.}

\item{value}{Atomic scalar (length 1, non-NA).}

\item{overwrite}{If \code{FALSE} (default) error when the scalar already
exists; if \code{TRUE} replace.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
Set a scalar.
}
\examples{
d <- memory_daf()
set_scalar(d, "organism", "human")
get_scalar(d, "organism")
}
