% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSetProjectName.R
\name{SetProjectName}
\alias{SetProjectName}
\title{Sets the internal name used by dafR to reference the project}
\usage{
SetProjectName(projectName)
}
\description{
This function allows you to define the project name you want instead of leaving that to the automation of GetProjectName. If there's an existing project name, it will be substituted by this one, so keep in mind that some expected behaviours, for instance, restoring data from S3 may not be exactly what you'd expect.

If you wish to manually set the project name do so before using any function that fetchs it using GetProjectName.

The functions that use GetProjectName are:
- RedshiftQuery()
}
\author{
Ricardo Vladimiro
}
