% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_daf}
\alias{setup_daf}
\title{Set up of the Julia environment needed for DataAxesFormats and TanayLabUtilities}
\usage{
setup_daf(
  pkg_check = TRUE,
  seed = NULL,
  env_path = getwd(),
  installJulia = FALSE,
  force_dev = TRUE,
  level = "Warn",
  show_time = TRUE,
  show_module = TRUE,
  show_location = FALSE,
  ...
)
}
\arguments{
\item{pkg_check}{(Default=TRUE) Check whether needed Julia packages
are installed}

\item{seed}{Seed to be used.}

\item{env_path}{The path to were the Julia environment should be created.
By default, this is the current working directory.}

\item{installJulia}{(Default=TRUE) Whether to install Julia}

\item{force_dev}{(Default=FALSE) Whether to force dev versions of packages}

\item{level}{Log level, one of "Debug", "Info", "Warn", "Error", or an integer (default: "Warn")}

\item{show_time}{Whether to show timestamp (default: TRUE)}

\item{show_module}{Whether to show module name (default: TRUE)}

\item{show_location}{Whether to show source location (default: FALSE)}

\item{...}{Other parameters passed on to \code{\link[JuliaCall]{julia_setup}}}
}
\value{
No return value, called for side effects.
}
\description{
This will set up a new Julia environment in the current working
directory or another folder if provided. This environment will
then be set with all Julia dependencies needed.
}
\examples{
\dontrun{
# Install from default URLs
setup_daf()

# Time consuming and requires Julia
setup_daf(installJulia = TRUE, seed = 60427)

# Install from latest github versions
setup_daf(
    installJulia = TRUE,
    custom_urls = c(
        "https://github.com/tanaylab/DataAxesFormats.jl",
        "https://github.com/tanaylab/TanayLabUtilities.jl"
    )
)
}
}
