% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoleShout.R
\name{Shout}
\alias{Shout}
\title{Shout something in the console}
\usage{
Shout(...)
}
\arguments{
\item{...}{Variables to be sent to the console}
}
\description{
Shout is a cat to stderr, meaning it will print like an error but without halting execution. It will output a sentence built from the arguments passed. A carriage return is added for ease of use.
}

