% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperSourceDirectory.R
\name{SourceDirectory}
\alias{SourceDirectory}
\title{Source all R files of a directory}
\usage{
SourceDirectory(directoryPath, ...)
}
\description{
Taken from the source help file and slightly changed to fit dafR
}
\author{
R Core Team and Ricardo Vladimiro
}
