% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{SquareMaskRow}
\alias{SquareMaskRow}
\title{SquareMaskRow query operation}
\usage{
SquareMaskRow(value, ...)
}
\arguments{
\item{value}{String specifying the value}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Similar to \code{MaskSlice} but is used when the mask matrix is square and we'd like to use a row as a mask. See the
Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.SquareMaskRow}{documentation} for
details.
}
