% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Std}
\alias{Std}
\title{Reduction query operation: standard deviation.}
\usage{
Std(type = NULL, ...)
}
\arguments{
\item{type}{Optional output type (character scalar) or a piped
\link{DafrQuery}.}

\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{\% Std} query fragment.
}
\examples{
Std()
}
\seealso{
\code{\link[=Var]{Var()}}, \code{\link[=StdN]{StdN()}}
}
