% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daf.R
\name{[.Daf}
\alias{[.Daf}
\title{Extract results from a Daf object using a query}
\usage{
\method{[}{Daf}(x, i, ...)
}
\arguments{
\item{x}{A Daf object}

\item{i}{A query string or object}

\item{...}{Ignored. Present for compatibility with the \code{[} generic.}
}
\value{
The result of the query
}
\description{
Extract results from a Daf object using a query
}
\details{
The expression \code{daf[query]} is equivalent to \code{get_query(daf, query, cache = FALSE)}.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Operations.get_query}{documentation} for details.
}
