% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Sum}
\alias{Sum}
\title{Reduction query operation: sum.}
\usage{
Sum(type = NULL, ...)
}
\arguments{
\item{type}{Optional output type (character scalar) or a piped
\link{DafrQuery}.}

\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{\% Sum} query fragment (optionally \verb{\% Sum type: <T>}).
}
\examples{
Sum()
Sum(type = "Float64")
}
\seealso{
\code{\link[=Mean]{Mean()}}, \code{\link[=Max]{Max()}}
}
