% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_import.R
\name{use_default_julia_environment}
\alias{use_default_julia_environment}
\title{Use default or custom Julia environment}
\usage{
use_default_julia_environment(env_path = "default")
}
\arguments{
\item{env_path}{Either "default" to use the default Julia environment, or a path to a custom environment}
}
\value{
No return value, called for side effects.
}
\description{
Force JuliaCall to use your specified Julia environment instead of creating a new one.
By default JuliaCall creates its own separate independent environment, which means
that it re-downloads and re-installs all the dependency packages there.
This function allows using the default Julia environment or a custom one.
}
