% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{Var}
\alias{Var}
\title{Reduction query operation: variance.}
\usage{
Var(type = NULL, ...)
}
\arguments{
\item{type}{Optional output type (character scalar) or a piped
\link{DafrQuery}.}

\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{\% Var} query fragment.
}
\examples{
Var()
}
\seealso{
\code{\link[=Std]{Std()}}, \code{\link[=VarN]{VarN()}}
}
