% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{vector_contract}
\alias{vector_contract}
\title{Create a vector contract specification}
\usage{
vector_contract(axis, name, expectation, dtype, description)
}
\arguments{
\item{axis}{Axis name the vector belongs to}

\item{name}{Vector name}

\item{expectation}{One of RequiredInput, OptionalInput, GuaranteedOutput, OptionalOutput}

\item{dtype}{Data type (e.g., "character", "numeric", "integer", "logical")}

\item{description}{Human-readable description}
}
\value{
Vector specification list
}
\description{
Create a vector contract specification
}
