% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{vectors_set}
\alias{vectors_set}
\title{Names of vectors on an axis, sorted.}
\usage{
vectors_set(daf, axis)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{axis}{Axis name.}
}
\value{
Character vector.
}
\description{
Names of vectors on an axis, sorted.
}
\examples{
d <- example_cells_daf()
vectors_set(d, "cell")
}
