% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\docType{data}
\name{VIEW_ALL_AXES}
\alias{VIEW_ALL_AXES}
\title{Default view spec: expose all axes as-is.}
\format{
An object of class \code{list} of length 1.
}
\usage{
VIEW_ALL_AXES
}
\description{
A pre-built \code{axes} override list for \code{viewer()} that exposes every axis of
the base daf unchanged. Equivalent to \code{list(list(ALL_AXES, "="))}.
}
\examples{
d <- example_cells_daf()
v <- viewer(d, axes = list(VIEW_ALL_AXES))
axes_set(v)
}
\seealso{
\code{\link[=viewer]{viewer()}}, \link{ALL_AXES}
}
\keyword{datasets}
