% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\docType{data}
\name{VIEW_ALL_MATRICES}
\alias{VIEW_ALL_MATRICES}
\title{Default view spec: expose all matrices as-is.}
\format{
An object of class \code{list} of length 1.
}
\usage{
VIEW_ALL_MATRICES
}
\description{
A pre-built \code{data} item for \code{viewer()} that exposes every matrix of the base
daf unchanged.
}
\examples{
d <- example_cells_daf()
v <- viewer(d, data = VIEW_ALL_MATRICES)
matrices_set(v, "cell", "gene")
}
\seealso{
\code{\link[=viewer]{viewer()}}, \link{ALL_MATRICES}
}
\keyword{datasets}
