% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\docType{data}
\name{VIEW_ALL_VECTORS}
\alias{VIEW_ALL_VECTORS}
\title{Default view spec: expose all vectors as-is.}
\format{
An object of class \code{list} of length 1.
}
\usage{
VIEW_ALL_VECTORS
}
\description{
A pre-built \code{data} item for \code{viewer()} that exposes every vector of every
axis of the base daf unchanged.
}
\examples{
d <- example_cells_daf()
v <- viewer(d, data = VIEW_ALL_VECTORS)
vectors_set(v, "cell")
}
\seealso{
\code{\link[=viewer]{viewer()}}, \link{ALL_VECTORS}
}
\keyword{datasets}
