% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoleWindmill.R
\name{Windmill}
\alias{Windmill}
\title{Update something in the terminal.}
\usage{
Windmill(...)
}
\arguments{
\item{...}{Variables to be sent to the console}
}
\description{
Say something in the same line while updating a windmill text animation. Serves to give visual feedback.
}

