% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{XorMask}
\alias{XorMask}
\title{Xor-mask query operation.}
\usage{
XorMask(value, ...)
}
\arguments{
\item{value}{Property name (character scalar), or a piped
\link{DafrQuery}.}

\item{...}{Optional piped \link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{^ <property>} query fragment that chains a logical XOR
condition into an open mask subquery.
}
\examples{
XorMask("type")
BeginMask("a") |> XorMask("b") |> EndMask()
}
\seealso{
\code{\link[=XorNegatedMask]{XorNegatedMask()}}, \code{\link[=AndMask]{AndMask()}}, \code{\link[=OrMask]{OrMask()}}, \code{\link[=BeginMask]{BeginMask()}}, \code{\link[=EndMask]{EndMask()}}
}
