% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builders_exports.R
\name{XorNegatedMask}
\alias{XorNegatedMask}
\title{Xor-negated-mask query operation.}
\usage{
XorNegatedMask(value, ...)
}
\arguments{
\item{...}{Optional piped \link{DafrQuery}.}

\item{property}{Property name (character scalar), or a piped
\link{DafrQuery}.}
}
\value{
A \link{DafrQuery}.
}
\description{
Builds a \verb{^ ! <property>} query fragment that chains a logical XOR NOT
condition into an open mask subquery.
}
\examples{
XorNegatedMask("type")
BeginMask("a") |> XorNegatedMask("b") |> EndMask()
}
\seealso{
\code{\link[=XorMask]{XorMask()}}, \code{\link[=AndNegatedMask]{AndNegatedMask()}}, \code{\link[=OrNegatedMask]{OrNegatedMask()}}, \code{\link[=BeginMask]{BeginMask()}}, \code{\link[=EndMask]{EndMask()}}
}
