% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{XorNot}
\alias{XorNot}
\title{XorNot query operation}
\usage{
XorNot(property, ...)
}
\arguments{
\item{property}{String specifying the property}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Same as \code{Xor} but use the inverse of the mask. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAccessFormats.Queries.XorNot}{documentation} for details.
}
