% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Abs}
\alias{Abs}
\title{Abs query operation}
\usage{
Abs(type = NULL, ...)
}
\arguments{
\item{type}{Optional result type (e.g., "Float64"). If NULL, the default type is used.}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Element-wise operation that converts every element to its absolute value. This operation preserves
the shape of the data (scalar, vector, or matrix) but changes the values. By default, the output
data type is the unsigned version of the input data type. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/operations.html#DataAxesFormats.Operations.Abs}{documentation} for
details.
}
