% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{add_axis}
\alias{add_axis}
\title{Add a new axis.}
\usage{
add_axis(daf, axis, entries)
}
\arguments{
\item{daf}{A \code{DafWriter}.}

\item{axis}{Axis name.}

\item{entries}{Unique, non-NA, non-empty character vector of entry names.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
Add a new axis.
}
\examples{
# Mirrors writers.jl jldoctest at line 161.
d <- example_cells_daf()
has_axis(d, "block")                  # FALSE
add_axis(d, "block", c("B1", "B2"))
has_axis(d, "block")                  # TRUE
}
