% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{add_axis}
\alias{add_axis}
\title{Add axis to a Daf object}
\usage{
add_axis(daf, axis, entries, overwrite = FALSE)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the new axis}

\item{entries}{Vector of entry names (must be unique within the axis)}

\item{overwrite}{Whether to overwrite if axis already exists (FALSE by default)}
}
\value{
The Daf object (invisibly, for chaining operations)
}
\description{
Creates a new axis with the specified name and entries in the Daf data set.
}
\details{
This function creates a new axis with the specified unique entry names.
If the axis already exists and \code{overwrite} is FALSE, an error will be raised.
Entry names must be unique within the axis.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/writers.html#DataAxesFormats.Writers.add_axis!}{documentation} for details.
}
