% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\docType{data}
\name{ALL_AXES}
\alias{ALL_AXES}
\title{Wildcard for all axes.}
\format{
An object of class \code{character} of length 1.
}
\usage{
ALL_AXES
}
\description{
Wildcard for all axes.
}
\examples{
d <- example_cells_daf()
v <- viewer(d, axes = list(list(ALL_AXES, "=")))
axes_set(v)
}
\keyword{datasets}
