% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\docType{data}
\name{ALL_AXES}
\alias{ALL_AXES}
\title{All axes specifier}
\format{
Character string
}
\usage{
ALL_AXES
}
\description{
A pair to use in the \code{axes} parameter of \code{viewer} to specify all the base data axes.
See the Julia documentation for details.
}
\references{
\url{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/views.html#DataAxesFormats.Views.ALL_AXES}
}
\keyword{datasets}
