% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\docType{data}
\name{ALL_MATRICES}
\alias{ALL_MATRICES}
\title{Wildcard for all matrices.}
\format{
An object of class \code{character} of length 3.
}
\usage{
ALL_MATRICES
}
\description{
Wildcard for all matrices.
}
\keyword{datasets}
