% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\docType{data}
\name{ALL_MATRICES}
\alias{ALL_MATRICES}
\title{Wildcard for all matrices.}
\format{
An object of class \code{character} of length 3.
}
\usage{
ALL_MATRICES
}
\description{
Wildcard for all matrices.
}
\examples{
d <- example_cells_daf()
v <- viewer(d, data = list(list(ALL_MATRICES, "=")))
matrices_set(v, "cell", "gene")
}
\keyword{datasets}
