% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\docType{data}
\name{ALL_MATRICES}
\alias{ALL_MATRICES}
\title{All matrices specifier}
\format{
Character vector of length 3
}
\usage{
ALL_MATRICES
}
\description{
A key to use in the \code{data} parameter of \code{viewer} to specify all the matrices of the exposed axes.
See the Julia documentation for details.
}
\references{
\url{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/views.html#DataAccessFormats.Views.ALL_MATRICES}
}
\keyword{datasets}
