% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\docType{data}
\name{ALL_SCALARS}
\alias{ALL_SCALARS}
\title{Wildcard for all scalars.}
\format{
An object of class \code{character} of length 1.
}
\usage{
ALL_SCALARS
}
\description{
Wildcard for all scalars.
}
\keyword{datasets}
