% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\docType{data}
\name{ALL_SCALARS}
\alias{ALL_SCALARS}
\title{All scalars specifier}
\format{
Character string
}
\usage{
ALL_SCALARS
}
\description{
A key to use in the \code{data} parameter of \code{viewer} to specify all the base data scalars.
See the Julia documentation for details.
}
\references{
\url{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/views.html#DataAxesFormats.Views.ALL_SCALARS}
}
\keyword{datasets}
