% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\docType{data}
\name{ALL_VECTORS}
\alias{ALL_VECTORS}
\title{Wildcard for all vectors.}
\format{
An object of class \code{character} of length 2.
}
\usage{
ALL_VECTORS
}
\description{
Wildcard for all vectors.
}
\keyword{datasets}
