% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\docType{data}
\name{ALL_VECTORS}
\alias{ALL_VECTORS}
\title{All vectors specifier}
\format{
Character vector of length 2
}
\usage{
ALL_VECTORS
}
\description{
A key to use in the \code{data} parameter of \code{viewer} to specify all the vectors of the exposed axes.
See the Julia documentation for details.
}
\references{
\url{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/views.html#DataAxesFormats.Views.ALL_VECTORS}
}
\keyword{datasets}
