% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata_facade.R
\name{as_anndata}
\alias{as_anndata}
\title{One-shot factory for a \link{DafAnnData} facade.}
\usage{
as_anndata(daf, obs_axis = NULL, var_axis = NULL, x_name = "UMIs")
}
\arguments{
\item{daf}{A \link{DafReader}.}

\item{obs_axis}{Axis name for observations. Defaults auto-detect:
\code{"cell"} then \code{"metacell"}.}

\item{var_axis}{Axis name for variables. Defaults to \code{"gene"}.}

\item{x_name}{Matrix name for \code{X}. Default \code{"UMIs"}.}
}
\value{
A \link{DafAnnData} instance.
}
\description{
One-shot factory for a \link{DafAnnData} facade.
}
\examples{
d <- memory_daf()
add_axis(d, "cell", c("c1", "c2"))
add_axis(d, "gene", c("g1", "g2"))
set_matrix(d, "cell", "gene", "UMIs", matrix(1:4, 2, 2))
ann <- as_anndata(d)
}
\seealso{
\link{DafAnnData}
}
