% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{AsAxis}
\alias{AsAxis}
\title{AsAxis query operation}
\usage{
AsAxis(axis = NULL, ...)
}
\arguments{
\item{axis}{Optional string specifying the axis}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
There are three cases where we may want to take a vector property and consider each value to be the name of an entry
of some axis: \code{Fetch}, \code{CountBy} and \code{GroupBy}. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.AsAxis}{documentation} for
details.
}
