% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Axis}
\alias{Axis}
\title{Axis query operation}
\usage{
Axis(axis, ...)
}
\arguments{
\item{axis}{String specifying the axis name}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
A query operation for specifying a result axis in a query sequence.
This is typically the first operation in a query sequence and determines which axis
the query will operate on. It sets the context for subsequent operations in the query.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.Axis}{documentation} for details.
}
