% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{axis_contract}
\alias{axis_contract}
\title{Create an axis contract specification}
\usage{
axis_contract(name, expectation, description)
}
\arguments{
\item{name}{Axis name}

\item{expectation}{One of RequiredInput, OptionalInput, GuaranteedOutput, OptionalOutput}

\item{description}{Human-readable description}
}
\value{
Axis specification list
}
\description{
Create an axis contract specification
}
