% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{axis_dict}
\alias{axis_dict}
\title{Entry-name to 1-based-index hash for an axis.}
\usage{
axis_dict(daf, axis)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{axis}{Axis name (character scalar).}
}
\value{
An environment mapping entry names to integer positions.
}
\description{
Entry-name to 1-based-index hash for an axis.
}
