% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{axis_indices}
\alias{axis_indices}
\title{Get indices of entries in an axis}
\usage{
axis_indices(daf, axis, entries, allow_empty = FALSE)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis}

\item{entries}{Character vector of entry names to look up}

\item{allow_empty}{Whether to allow empty entries (return -1 for empty strings if TRUE)}
}
\value{
A vector of 1-based indices corresponding to the entries
}
\description{
Returns the integer indices for specified entry names along an axis.
}
\details{
This function maps names to their position indices along the axis.
If \code{allow_empty} is TRUE, empty strings are converted to index -1.
Indices in R are 1-based (first element has index 1), consistent with R conventions.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.axis_indices}{documentation} for details.
}
