% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{axis_length}
\alias{axis_length}
\title{Get length of an axis in a Daf object}
\usage{
axis_length(daf, axis)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis}
}
\value{
Length (number of entries) of the axis
}
\description{
Returns the number of entries along the specified axis in the Daf data set.
}
\details{
The axis length corresponds to the size of vector properties for this axis
and to one of the dimensions of matrix properties involving this axis.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.axis_length}{documentation} for details.
}
