% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{axis_vector}
\alias{axis_vector}
\title{Get vector of axis entries from a Daf object}
\usage{
axis_vector(daf, axis, null_if_missing = FALSE)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis}

\item{null_if_missing}{Whether to return NULL if the axis doesn't exist}
}
\value{
A character vector of axis entry names
}
\description{
Returns a vector of the unique names for all entries of the specified axis.
}
\details{
Axis entries provide names for each position along an axis, such as gene names
for a "gene" axis or cell barcodes for a "cell" axis. These entry names can be used
to look up specific indices using the \code{axis_indices()} function.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.axis_vector}{documentation} for details.
}
