% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{axis_version_counter}
\alias{axis_version_counter}
\title{Get axis version counter}
\usage{
axis_version_counter(daf, axis)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis}
}
\value{
An integer version counter
}
\description{
Returns the version counter for an axis, which is incremented when the axis is modified.
This is useful for cache invalidation.
}
\details{
The version counter is incremented whenever the axis entries are modified.
This can be used to detect changes and invalidate caches.
}
