% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_daf.R
\name{chain_reader}
\alias{chain_reader}
\title{Create a read-only chain of DafReaders.}
\usage{
chain_reader(dafs, name = NULL)
}
\arguments{
\item{dafs}{Ordered list of \code{DafReader}s. Later entries override earlier
entries on read.}

\item{name}{Optional chain name; defaults to \code{paste(names, collapse = ";")}.}
}
\value{
A \code{ReadOnlyChainDaf}.
}
\description{
Create a read-only chain of DafReaders.
}
