% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{chain_reader}
\alias{chain_reader}
\title{Create a read-only chain wrapper of DafReader objects}
\usage{
chain_reader(dsets, name = NULL)
}
\arguments{
\item{dsets}{List of Daf objects to chain}

\item{name}{Optional name for the chained Daf object}
}
\value{
A read-only Daf object chaining the input Daf objects
}
\description{
This function creates a read-only chain wrapper of DafReader objects, presenting them as a single DafReader.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/chains.html#DataAxesFormats.Chains.chain_reader}{documentation} for details.
}
