% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{chain_writer}
\alias{chain_writer}
\title{Create a writable chain wrapper of DafReader objects}
\usage{
chain_writer(dsets, name = NULL)
}
\arguments{
\item{dsets}{List of Daf objects to chain}

\item{name}{Optional name for the chained Daf object}
}
\value{
A writable Daf object chaining the input Daf objects
}
\description{
This function creates a chain wrapper for a chain of DafReader data, presenting them as a single DafWriter.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/chains.html#DataAxesFormats.Chains.chain_writer}{documentation} for details.
}
