% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Clamp}
\alias{Clamp}
\title{Clamp query operation}
\usage{
Clamp(min = NULL, max = NULL, ...)
}
\arguments{
\item{min}{Minimum value; values less than this will be set to this value}

\item{max}{Maximum value; values greater than this will be set to this value}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Element-wise operation that converts every element to a value inside a range.
This operation preserves the shape of the data (scalar, vector, or matrix) but changes
the values, setting each value to be within the specified range. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/operations.html#DataAxesFormats.Operations.Clamp}{documentation} for
details.
}
