% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate multiple Daf data sets along some axis}
\usage{
concatenate(
  destination,
  axis,
  sources,
  names = NULL,
  dataset_axis = "dataset",
  dataset_property = TRUE,
  prefix = FALSE,
  prefixed = NULL,
  empty = NULL,
  sparse_if_saves_storage_fraction = 0.1,
  merge = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{destination}{A Daf object to write the concatenated result to}

\item{axis}{Name of the axis or axes to concatenate along}

\item{sources}{A list of Daf objects to concatenate}

\item{names}{Optional vector of names for the sources}

\item{dataset_axis}{Optional name for the dataset axis}

\item{dataset_property}{Whether to add a dataset property}

\item{prefix}{Whether to prefix the names of properties in the source data sets}

\item{prefixed}{Optional set of names that are already prefixed and should be left alone}

\item{empty}{Value to use for filling in missing data}

\item{sparse_if_saves_storage_fraction}{Fraction of storage savings required to use sparse matrices}

\item{merge}{Optional named list mapping property keys to merge actions ("SkipProperty", "LastValue", or "CollectAxis")}

\item{overwrite}{Whether to overwrite existing data}
}
\value{
The destination Daf object (invisibly)
}
\description{
Concatenate multiple Daf data sets along some axis
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/concat.html}{documentation} for details.
}
