% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Convert}
\alias{Convert}
\title{Convert query operation}
\usage{
Convert(type, ...)
}
\arguments{
\item{type}{Type to convert to (e.g., "Int32", "Float64")}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Element-wise operation that converts every element to a given data type.
This operation preserves the shape of the data (scalar, vector, or matrix) but
changes the data type of the elements. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/operations.html#DataAxesFormats.Operations.Convert}{documentation} for
details.
}
