% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copies.R
\name{copy_all}
\alias{copy_all}
\title{Copy all content from source to destination}
\usage{
copy_all(
  destination,
  source,
  empty = NULL,
  overwrite = FALSE,
  relayout = TRUE,
  types = NULL,
  insist = TRUE
)
}
\arguments{
\item{destination}{A Daf object to copy to}

\item{source}{A Daf object to copy from}

\item{empty}{A named list mapping data keys to values for filling missing data}

\item{overwrite}{Whether to overwrite if data already exists}

\item{relayout}{Whether to allow relayout}

\item{types}{Optional named list mapping data keys to types for conversion. If NULL, the original types are preserved.}

\item{insist}{Whether to fail if data doesn't exist (TRUE by default)}
}
\value{
The destination Daf object (invisibly)
}
\description{
Copy all content from source to destination
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/copies.html#DataAxesFormats.Copies.copy_all!}{documentation} for details.
}
