% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copies.R
\name{copy_axis}
\alias{copy_axis}
\title{Copy an axis from source to destination}
\usage{
copy_axis(
  destination,
  source,
  axis,
  rename = NULL,
  default = NULL,
  overwrite = FALSE,
  insist = TRUE
)
}
\arguments{
\item{destination}{A Daf object to copy to}

\item{source}{A Daf object to copy from}

\item{axis}{Name of the axis to copy}

\item{rename}{Optional new name for the axis in the destination}

\item{default}{Default value if axis doesn't exist}

\item{overwrite}{Whether to overwrite if axis already exists}

\item{insist}{Whether to fail if the axis doesn't exist}
}
\value{
The destination Daf object (invisibly)
}
\description{
Copy an axis from source to destination
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/copies.html#DataAxesFormats.Copies.copy_axis!}{documentation} for details.
}
