% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copies.R
\name{copy_matrix}
\alias{copy_matrix}
\title{Copy a matrix from source to destination}
\usage{
copy_matrix(
  destination,
  source,
  rows_axis,
  columns_axis,
  name,
  rows_reaxis = NULL,
  columns_reaxis = NULL,
  rename = NULL,
  default,
  empty = NULL,
  relayout = TRUE,
  overwrite = FALSE,
  eltype = NULL,
  bestify = FALSE,
  min_sparse_saving_fraction = NULL,
  insist = TRUE
)
}
\arguments{
\item{destination}{A Daf object to copy to}

\item{source}{A Daf object to copy from}

\item{rows_axis}{Name of rows axis}

\item{columns_axis}{Name of columns axis}

\item{name}{Name of the matrix to copy}

\item{rows_reaxis}{Optional different rows axis name in the destination}

\item{columns_reaxis}{Optional different columns axis name in the destination}

\item{rename}{Optional new name for the matrix in the destination}

\item{default}{Default value if matrix doesn't exist}

\item{empty}{Value to use for filling in missing data}

\item{relayout}{Whether to allow relayout}

\item{overwrite}{Whether to overwrite if matrix already exists}

\item{eltype}{Optional element type to convert to (e.g., "Float64", "Int32"). If NULL, the original type is preserved.}

\item{bestify}{Whether to bestify the matrix storage (FALSE by default)}

\item{min_sparse_saving_fraction}{Optional minimum sparse saving fraction. If NULL, the default is used.}

\item{insist}{Whether to fail if the matrix doesn't exist (TRUE by default)}
}
\value{
The destination Daf object (invisibly)
}
\description{
Copy a matrix from source to destination
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/copies.html#DataAxesFormats.Copies.copy_matrix!}{documentation} for details.
}
