% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copies.R
\name{copy_scalar}
\alias{copy_scalar}
\title{Copy a scalar from source to destination}
\usage{
copy_scalar(
  destination,
  source,
  name,
  rename = NULL,
  default = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{destination}{A Daf object to copy to}

\item{source}{A Daf object to copy from}

\item{name}{Name of the scalar to copy}

\item{rename}{Optional new name for the scalar in the destination}

\item{default}{Default value if scalar doesn't exist}

\item{overwrite}{Whether to overwrite if scalar already exists}
}
\value{
The destination Daf object (invisibly)
}
\description{
Copy a scalar from source to destination
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/copies.html#DataAxesFormats.Copies.copy_scalar!}{documentation} for details.
}
