% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copies.R
\name{copy_vector}
\alias{copy_vector}
\title{Copy a vector from source to destination}
\usage{
copy_vector(
  destination,
  source,
  axis,
  name,
  reaxis = NULL,
  rename = NULL,
  default = NULL,
  empty = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{destination}{A Daf object to copy to}

\item{source}{A Daf object to copy from}

\item{axis}{Name of the axis}

\item{name}{Name of the vector to copy}

\item{reaxis}{Optional different axis name in the destination}

\item{rename}{Optional new name for the vector in the destination}

\item{default}{Default value if vector doesn't exist}

\item{empty}{Value to use for filling in missing data}

\item{overwrite}{Whether to overwrite if vector already exists}
}
\value{
The destination Daf object (invisibly)
}
\description{
Copy a vector from source to destination
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/copies.html#DataAxesFormats.Copies.copy_vector!}{documentation} for details.
}
