% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copies.R
\name{copy_vector}
\alias{copy_vector}
\title{Copy a vector from source to destination}
\usage{
copy_vector(
  destination,
  source,
  axis,
  name,
  reaxis = NULL,
  rename = NULL,
  default,
  empty = NULL,
  overwrite = FALSE,
  eltype = NULL,
  bestify = FALSE,
  min_sparse_saving_fraction = NULL,
  insist = TRUE
)
}
\arguments{
\item{destination}{A Daf object to copy to}

\item{source}{A Daf object to copy from}

\item{axis}{Name of the axis}

\item{name}{Name of the vector to copy}

\item{reaxis}{Optional different axis name in the destination}

\item{rename}{Optional new name for the vector in the destination}

\item{default}{Default value if vector doesn't exist}

\item{empty}{Value to use for filling in missing data}

\item{overwrite}{Whether to overwrite if vector already exists}

\item{eltype}{Optional element type to convert to (e.g., "Float64", "Int32"). If NULL, the original type is preserved.}

\item{bestify}{Whether to bestify the vector storage (FALSE by default)}

\item{min_sparse_saving_fraction}{Optional minimum sparse saving fraction. If NULL, the default is used.}

\item{insist}{Whether to fail if the vector doesn't exist (TRUE by default)}
}
\value{
The destination Daf object (invisibly)
}
\description{
Copy a vector from source to destination
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/copies.html#DataAxesFormats.Copies.copy_vector!}{documentation} for details.
}
