% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{CountBy}
\alias{CountBy}
\title{CountBy query operation}
\usage{
CountBy(property, ...)
}
\arguments{
\item{property}{String specifying the property}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
A query operation that generates a matrix of counts of combinations of pairs of values for the same entries of an
axis. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAccessFormats.Queries.CountBy}{documentation} for
details.
}
