% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{create_contract}
\alias{create_contract}
\title{Create a Contract specification}
\usage{
create_contract(axes = list(), data = list(), is_relaxed = FALSE)
}
\arguments{
\item{axes}{List of axis specifications (see axis_contract)}

\item{data}{List of data specifications (see vector_contract, matrix_contract, scalar_contract)}

\item{is_relaxed}{If TRUE, allows additional inputs/outputs not in contract}
}
\value{
A contract object that can be used with verify_contract
}
\description{
Creates a contract specification that can be used to validate DAF data.
This corresponds to Julia's DataAxesFormats.Contracts.Contract.
}
\examples{
\dontrun{
contract <- create_contract(
    axes = list(
        axis_contract("metacell", RequiredInput, "Metacell identifiers"),
        axis_contract("gene", RequiredInput, "Gene identifiers")
    ),
    data = list(
        vector_contract(
            "metacell", "type", RequiredInput, "character",
            "Cell type per metacell"
        ),
        matrix_contract(
            "metacell", "gene", "UMIs", RequiredInput, "numeric",
            "UMI counts matrix"
        )
    )
)
}
}
