% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daf.R
\name{Daf}
\alias{Daf}
\title{Daf (Data Axes Format) S3 object}
\usage{
Daf(jl_obj)
}
\arguments{
\item{jl_obj}{The Julia DafReader object to wrap}
}
\value{
A Daf S3 object
}
\description{
An S3 object that wraps a Julia DataAxesFormats object and provides access to its methods.
}
\details{
See the Julia documentation \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/index.html}{here},
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/formats.html#Read-API}{here}
and \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/formats.html#Write-API}{here} for details.
}
