% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata_format.R
\name{daf_as_h5ad}
\alias{daf_as_h5ad}
\title{Convert Daf object to h5ad file}
\usage{
daf_as_h5ad(
  daf,
  h5ad,
  obs_is = NULL,
  var_is = NULL,
  X_is = NULL,
  X_eltype = NULL
)
}
\arguments{
\item{daf}{A Daf object}

\item{h5ad}{Path where the h5ad file will be written}

\item{obs_is}{Optional name for the observation axis}

\item{var_is}{Optional name for the variable axis}

\item{X_is}{Optional name for the main matrix}

\item{X_eltype}{Optional element type for the X matrix (e.g., "Float32"). If NULL, the original type is preserved.}
}
\value{
Invisibly returns the input Daf object
}
\description{
View the Daf data set as AnnData and save it to an h5ad file
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/anndata_format.html#DataAxesFormats.AnnDataFormat.daf_as_anndata}{documentation} for details.
Note this just creates the h5ad file. The Julia (Muon.jl) AnnData object is not returned or exposed to R.
}
