% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{delete_axis}
\alias{delete_axis}
\title{Delete axis from a Daf object}
\usage{
delete_axis(daf, axis, must_exist = TRUE)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis to delete}

\item{must_exist}{Whether to error if axis doesn't exist (TRUE by default)}
}
\value{
The Daf object (invisibly, for chaining operations)
}
\description{
Removes an axis and all its associated data from the Daf data set.
}
\details{
This function deletes an axis and all vector and matrix properties
associated with it. If \code{must_exist} is TRUE and the axis doesn't exist,
an error will be raised.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/writers.html#DataAxesFormats.Writers.delete_axis!}{documentation} for details.
}
