% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{delete_matrix}
\alias{delete_matrix}
\title{Delete a matrix.}
\usage{
delete_matrix(daf, rows_axis, cols_axis, name, must_exist = TRUE)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{rows_axis}{Row-axis name.}

\item{cols_axis}{Column-axis name.}

\item{name}{Matrix name.}

\item{must_exist}{See \code{delete_axis}.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
Delete a matrix.
}
