% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{delete_matrix}
\alias{delete_matrix}
\title{Delete matrix from a Daf object}
\usage{
delete_matrix(daf, rows_axis, columns_axis, name, must_exist = TRUE)
}
\arguments{
\item{daf}{A Daf object}

\item{rows_axis}{Name of rows axis}

\item{columns_axis}{Name of columns axis}

\item{name}{Name of the matrix property to delete}

\item{must_exist}{Whether to error if matrix doesn't exist (TRUE by default)}
}
\value{
The Daf object (invisibly, for chaining operations)
}
\description{
Removes a matrix property with the specified name from the Daf data set.
}
\details{
If \code{must_exist} is TRUE and the matrix doesn't exist, an error will be raised.
Otherwise, the function will silently succeed even if the matrix doesn't exist.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/writers.html#DataAxesFormats.Writers.delete_matrix!}{documentation} for details.
}
