% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{delete_scalar}
\alias{delete_scalar}
\title{Delete scalar from a Daf object}
\usage{
delete_scalar(daf, name, must_exist = TRUE)
}
\arguments{
\item{daf}{A Daf object}

\item{name}{Name of the scalar property to delete}

\item{must_exist}{Whether to error if scalar doesn't exist (TRUE by default)}
}
\value{
The Daf object (invisibly, for chaining operations)
}
\description{
Removes a scalar property with the specified name from the Daf data set.
}
\details{
If \code{must_exist} is TRUE and the scalar doesn't exist, an error will be raised.
Otherwise, the function will silently succeed even if the scalar doesn't exist.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/writers.html#DataAxesFormats.Writers.delete_scalar!}{documentation} for details.
}
