% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{delete_vector}
\alias{delete_vector}
\title{Delete vector from a Daf object}
\usage{
delete_vector(daf, axis, name, must_exist = TRUE)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Axis name}

\item{name}{Name of the vector property to delete}

\item{must_exist}{Whether to error if vector doesn't exist (TRUE by default)}
}
\value{
The Daf object (invisibly, for chaining operations)
}
\description{
Removes a vector property with the specified name from the Daf data set.
}
\details{
If \code{must_exist} is TRUE and the vector doesn't exist, an error will be raised.
Otherwise, the function will silently succeed even if the vector doesn't exist.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/writers.html#DataAxesFormats.Writers.delete_vector!}{documentation} for details.
}
