% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daf.R
\name{description}
\alias{description}
\title{Get description of a Daf object}
\usage{
description(daf, cache = FALSE, deep = FALSE, tensors = TRUE)
}
\arguments{
\item{daf}{A Daf object}

\item{cache}{Whether to include cache information}

\item{deep}{Whether to include detailed information about nested data}

\item{tensors}{Whether to include tensor information (condensed list of tensor matrices)}
}
\value{
A string description of the Daf object
}
\description{
Get description of a Daf object
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.description}{documentation} for details.
}
