% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{empty_cache}
\alias{empty_cache}
\title{Empty cache in a Daf object}
\usage{
empty_cache(daf, clear = NULL, keep = NULL)
}
\arguments{
\item{daf}{A Daf object}

\item{clear}{Cache group to clear. Can be one of "MappedData", "MemoryData", or "QueryData".}

\item{keep}{Cache group to keep. Can be one of "MappedData", "MemoryData", or "QueryData".}
}
\value{
The Daf object
}
\description{
Empty cache in a Daf object
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Data.empty_cache!}{documentation} for details.
}
