% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{escape_value}
\alias{escape_value}
\title{Escape a value for use in a query string}
\usage{
escape_value(value)
}
\arguments{
\item{value}{A character string value to escape}
}
\value{
The escaped value as a character string
}
\description{
Escapes special characters in a value so it can be safely embedded in a query string.
This is needed when values contain characters that have special meaning in the query
syntax (e.g., spaces, quotes, backslashes).
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.escape_value}{documentation} for details.
}
