% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_cells_daf}
\alias{example_cells_daf}
\title{Load example cells data into a Daf object}
\usage{
example_cells_daf(name = "cells!")
}
\arguments{
\item{name}{Name for the Daf object (default: "cells!")}
}
\value{
A Daf object containing example cells data
}
\description{
Load example cells data into a Daf object
}
\details{
This function loads example cells data from the DataAxesFormats.jl package.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/example_data.html}{documentation} for details.
}
