% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_chain_daf}
\alias{example_chain_daf}
\title{Load example chain data into a Daf object}
\usage{
example_chain_daf(name = "chain!")
}
\arguments{
\item{name}{Name for the Daf object (default: "chain!")}
}
\value{
A Daf object containing chained example cells and metacells data
}
\description{
Load example chain data into a Daf object
}
\details{
This function creates a chain of example cells and metacells data.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/example_data.html}{documentation} for details.
}
