% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_metacells_daf}
\alias{example_metacells_daf}
\title{Load example metacells data into a Daf object}
\usage{
example_metacells_daf(name = "metacells!")
}
\arguments{
\item{name}{Name for the Daf object (default: "metacells!")}
}
\value{
A Daf object containing example metacells data
}
\description{
Load example metacells data into a Daf object
}
\details{
This function loads example metacells data from the DataAxesFormats.jl package.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/example_data.html}{documentation} for details.
}
