% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Fetch}
\alias{Fetch}
\title{Fetch query operation}
\usage{
Fetch(property, ...)
}
\arguments{
\item{property}{String specifying the property name}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
A query operation for fetching the value of a property from another axis, based on a vector property whose values
are entry names of the axis. See the
Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.Fetch}{documentation} for
details.
}
