% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{files_daf}
\alias{files_daf}
\title{Create a Daf object with file-based storage}
\usage{
files_daf(path, mode = "r", name = NULL)
}
\arguments{
\item{path}{Path to the files storage location}

\item{mode}{Mode to open the storage ("r" for read-only, "r+" for read-write)}

\item{name}{Optional name for the Daf object}
}
\value{
A Daf object with file-based storage
}
\description{
This function creates a Daf object that stores data in disk files. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/files_format.html}{documentation} for details.
}
