% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Fraction}
\alias{Fraction}
\title{Fraction query operation}
\usage{
Fraction(...)
}
\arguments{
\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Element-wise operation that converts every element to its fraction out of the total.
This operation preserves the shape of the data (scalar, vector, or matrix) but changes
each value to be the fraction of the total sum of all values in the data. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/operations.html#DataAxesFormats.Operations.Fraction}{documentation} for
details.
}
