% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_dataframe}
\alias{get_dataframe}
\alias{get_tidy}
\title{Get a dataframe from a Daf object}
\usage{
get_dataframe(daf, axis, columns = NULL, cache = FALSE)

get_tidy(daf, axis, columns = NULL, cache = FALSE, ...)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Axis name or query object}

\item{columns}{Vector of column specifications or named list/vector mapping column names to queries}

\item{cache}{Whether to cache the query results (FALSE by default)}

\item{...}{Additional arguments passed to \code{tidyr::pivot_longer}}
}
\value{
A data.frame containing the specified columns for the axis, with row names set to the axis entries.
If columns is NULL, all columns are returned with the "name" column removed if present.

For \code{get_tidy}, a tibble in long format with columns "name", "key", and "value".
The "name" column contains the axis entries, "key" contains the column names,
and "value" contains the corresponding values.
Note that if the types of the columns are not homogeneous, an error will be thrown.
Use the \code{values_transform} argument to transform the types of the values,
e.g. \code{values_transform = list(value = as.character)}.
}
\description{
Retrieves multiple vector properties for an axis as a dataframe.
}
\details{
This function allows retrieving multiple vectors for the same axis in a single operation.
The \code{columns} parameter can be a vector of vector names, or a named list mapping output column names
to vector names or query strings.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.get_frame}{documentation} for more details.
}
