% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_dataframe_query}
\alias{get_dataframe_query}
\title{Apply a query to a Daf object and return result as a data.frame}
\usage{
get_dataframe_query(daf = NULL, query = NULL, cache = TRUE)
}
\arguments{
\item{daf}{A Daf object}

\item{query}{Query string or object. Can be created using query operations such as
Axis(), Lookup(), IsGreater(), etc.
In order to support the use of pipe operators, the query can also be a Daf object and vice versa.}

\item{cache}{Whether to cache the query result}
}
\value{
A data.frame representation of the query result. If the query result is a matrix, row names and column names are the axis entries. If the query result is a scalar, a data.frame with one row and one column is returned. If the query result is a vector, a data.frame with a column "value" is returned.
}
\description{
Executes a query on a Daf object and returns the result formatted as a data.frame for
easier use in R analysis workflows. This is a convenience wrapper around get_query
that provides properly structured dataframes for different result types.
}
\seealso{
get_query, get_frame
}
