% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_frame}
\alias{get_frame}
\title{Get a dataframe from a Daf object (Julia-style)}
\usage{
get_frame(daf, axis, columns = NULL, cache = FALSE)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Axis name or query object}

\item{columns}{Vector of column specifications or named list / vector mapping column names to queries}

\item{cache}{Whether to cache the query results}
}
\value{
A data.frame containing the specified columns for the axis. If columns is NULL, all columns are returned, with an additional column "name" containing the axis entries.
}
\description{
Retrieves multiple vector properties for an axis as a dataframe, returning the raw
Julia-style result. For a more R-friendly version, see \code{get_dataframe()}.
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.get_frame}{documentation} for more details.
}
