% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_query}
\alias{get_query}
\title{Evaluate a query string against a daf reader.}
\usage{
get_query(daf, query_string)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{query_string}{A query string.}
}
\value{
A scalar, vector, matrix, names set, or NULL if missing.
}
\description{
Evaluate a query string against a daf reader.
}
