% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_query}
\alias{get_query}
\title{Apply a query to a Daf object}
\usage{
get_query(daf = NULL, query = NULL, cache = TRUE)
}
\arguments{
\item{daf}{A Daf object}

\item{query}{Query string or object. Can be created using query operations such as
Axis(), Lookup(), IsGreater(), etc.
In order to support the use of pipe operators, the query can also be a Daf object and vice versa, see examples below.}

\item{cache}{Whether to cache the query result}
}
\value{
The result of the query
}
\description{
Apply a query to a Daf object
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Operations.get_query}{documentation} for details.
}
\seealso{
Axis, Lookup, Names, IsEqual, IsGreater, IsLess, and other query operations.
}
