% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_scalar}
\alias{get_scalar}
\title{Get scalar value from a Daf object}
\usage{
get_scalar(daf, name, default = NULL)
}
\arguments{
\item{daf}{A Daf object}

\item{name}{Name of the scalar property to retrieve}

\item{default}{Default value to return if the scalar doesn't exist. If NULL, an error is thrown.}
}
\value{
The scalar value or default if the property is not found
}
\description{
Retrieves the value of a scalar property with the given name from the Daf data set.
}
\details{
Numeric scalars are returned as integers or doubles, regardless of the specific
data type they are stored as in the Daf data set.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.get_scalar}{documentation} for details.
}
