% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_vector}
\alias{get_vector}
\title{Get vector from a Daf object}
\usage{
get_vector(daf, axis, name, default = NULL)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Axis name}

\item{name}{Name of the vector property}

\item{default}{Default value if vector doesn't exist (NULL by default)}
}
\value{
A named vector containing the property values, with names set to the axis entry names,
or the default value if the property doesn't exist
}
\description{
Retrieves a vector property with the specified name for a given axis.
}
\details{
Vector properties store one-dimensional data along an axis, with one value
for each entry in the axis. If the vector doesn't exist and default is NA,
a vector of NAs with appropriate length is returned.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.get_vector}{documentation} for details.
}
