% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_names}
\alias{group_names}
\title{Generate names for groups of axis entries}
\usage{
group_names(daf, axis, entries_of_groups, prefix)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis}

\item{entries_of_groups}{A list of integer vectors, where each vector contains
the indices of entries belonging to a group}

\item{prefix}{A prefix to add to each generated name}
}
\value{
A character vector of unique group names
}
\description{
Given groups of axis entries (as indices), generates unique names for each group
based on the entry names using a hash.
}
\details{
This function generates deterministic names for groups based on their
member entries. Groups with the same members will always get the same name.
This is useful for creating reproducible group identifiers.
}
