% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{GroupBy}
\alias{GroupBy}
\title{GroupBy query operation}
\usage{
GroupBy(property, ...)
}
\arguments{
\item{property}{String specifying the property}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
A query operation that uses a (following) \code{ReductionOperation} to aggregate the values of each group of
values. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAccessFormats.Queries.GroupBY}{documentation} for
details.
}
