% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndata_format.R
\name{h5ad_as_daf}
\alias{h5ad_as_daf}
\title{Convert h5ad file to a Daf object}
\usage{
h5ad_as_daf(
  h5ad,
  name = NULL,
  obs_is = NULL,
  var_is = NULL,
  X_is = NULL,
  unsupported_handler = WARN_HANDLER
)
}
\arguments{
\item{h5ad}{Path to the h5ad file}

\item{name}{Optional name for the Daf object}

\item{obs_is}{Optional name for the observation axis}

\item{var_is}{Optional name for the variable axis}

\item{X_is}{Optional name for the main matrix}

\item{unsupported_handler}{How to handle unsupported features (one of IGNORE_HANDLER, WARN_HANDLER, or ERROR_HANDLER)}
}
\value{
A Daf object
}
\description{
View AnnData as a Daf data set, specifically using a MemoryDaf
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/anndata_format.html#anndata_as_daf}{documentation} for details.
Note that this function works with h5ad file paths, as the Julia AnnData object comes from the Muon.jl package
and is not compatible with R/python anndata implementations.
}
