% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{h5df}
\alias{h5df}
\title{Create a Daf object with HDF5-based storage}
\usage{
h5df(root, mode = "r", name = NULL)
}
\arguments{
\item{root}{Path to the HDF5 file, or a Julia HDF5 File or Group object}

\item{mode}{Mode to open the storage ("r" for read-only, "r+" for read-write)}

\item{name}{Optional name for the Daf object}
}
\value{
A Daf object with HDF5-based storage
}
\description{
This function creates a Daf object that stores data in an HDF5 disk file. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/h5df_format.html}{documentation} for details.
}
