% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{has_axis}
\alias{has_axis}
\title{Test whether an axis exists.}
\usage{
has_axis(daf, axis)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{axis}{Axis name (character scalar).}
}
\value{
Logical scalar.
}
\description{
Test whether an axis exists.
}
\examples{
d <- memory_daf()
add_axis(d, "cell", c("c1", "c2"))
has_axis(d, "cell")
has_axis(d, "gene")
}
