% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{has_axis}
\alias{has_axis}
\title{Check if an axis exists in a Daf object}
\usage{
has_axis(daf, axis)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis to check}
}
\value{
TRUE if the axis exists, FALSE otherwise
}
\description{
Determines whether an axis with the specified name exists in the Daf data set.
}
\details{
Axes are fundamental dimensions in a Daf data set along which vector and matrix
data are stored. Each axis has a collection of unique named entries.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.has_axis}{documentation} for details.
}
