% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{has_matrix}
\alias{has_matrix}
\title{Test whether a matrix exists for an axis pair.}
\usage{
has_matrix(daf, rows_axis, cols_axis, name)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{rows_axis}{Row-axis name.}

\item{cols_axis}{Column-axis name.}

\item{name}{Matrix name.}
}
\value{
Logical scalar.
}
\description{
Test whether a matrix exists for an axis pair.
}
