% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{has_matrix}
\alias{has_matrix}
\title{Test whether a matrix exists for an axis pair.}
\usage{
has_matrix(daf, rows_axis, columns_axis, name)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{rows_axis}{Row-axis name.}

\item{columns_axis}{Column-axis name.}

\item{name}{Matrix name.}
}
\value{
Logical scalar.
}
\description{
Test whether a matrix exists for an axis pair.
}
\examples{
d <- example_cells_daf()
has_matrix(d, "gene", "cell", "UMIs")
has_matrix(d, "cell", "gene", "UMIs")
}
