% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{has_matrix}
\alias{has_matrix}
\title{Check if a matrix exists in a Daf object}
\usage{
has_matrix(daf, rows_axis, columns_axis, name, relayout = TRUE)
}
\arguments{
\item{daf}{A Daf object}

\item{rows_axis}{Name of rows axis}

\item{columns_axis}{Name of columns axis}

\item{name}{Name of the matrix property}

\item{relayout}{Whether to check with flipped axes too (TRUE by default)}
}
\value{
TRUE if matrix exists, FALSE otherwise
}
\description{
Determines whether a matrix property with the specified name exists for the given axes.
}
\details{
Matrix properties store two-dimensional data along two axes.
If \code{relayout} is TRUE, this function will also check if the matrix exists with
the axes flipped (i.e., rows as columns and columns as rows).
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.has_matrix}{documentation} for details.
}
