% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{has_query}
\alias{has_query}
\title{Check if a query can be applied to a Daf object}
\usage{
has_query(daf, query)
}
\arguments{
\item{daf}{A Daf object}

\item{query}{Query string or object. Can be created using query operations such as
Axis(), Lookup(), IsGreater(), etc.}
}
\value{
TRUE if query can be applied, FALSE otherwise
}
\description{
Check if a query can be applied to a Daf object
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Operations.has_query}{documentation} for details.
}
\seealso{
Axis, Lookup, Names, IsEqual, IsGreater, IsLess, and other query operations.
}
