% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{has_scalar}
\alias{has_scalar}
\title{Check if a scalar exists in a Daf object}
\usage{
has_scalar(daf, name)
}
\arguments{
\item{daf}{A Daf object}

\item{name}{Name of the scalar property to check}
}
\value{
TRUE if scalar exists, FALSE otherwise
}
\description{
Determines whether a scalar property with the specified name exists in the Daf data set.
}
\details{
Scalar properties are global values associated with the entire Daf data set.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.has_scalar}{documentation} for details.
}
