% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{has_vector}
\alias{has_vector}
\title{Check if a vector exists in a Daf object}
\usage{
has_vector(daf, axis, name)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis}

\item{name}{Name of the vector property}
}
\value{
TRUE if vector exists, FALSE otherwise
}
\description{
Determines whether a vector property with the specified name exists for the given axis.
}
\details{
Vector properties store one-dimensional data along a specific axis.
Each entry in the axis has a corresponding value in the vector.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.has_vector}{documentation} for details.
}
