% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{IfMissing}
\alias{IfMissing}
\title{IfMissing query operation}
\usage{
IfMissing(missing_value, type = NULL, ...)
}
\arguments{
\item{missing_value}{Value to use when data is missing}

\item{type}{Optional type specification}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
A query operation providing a value to use if the data is missing some property. See the
Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.IfMissing}{documentation}
for details.
}
