% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{IfMissing}
\alias{IfMissing}
\title{IfMissing query operation}
\usage{
IfMissing(missing_value, type = NULL, ...)
}
\arguments{
\item{missing_value}{Value to use when data is missing the property}

\item{type}{Optional type specification for the missing value (e.g., "Int64", "Float64", "Bool")}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
A query operation providing a default value to use if the data is missing some property.
This is useful when querying for properties that might not exist for all entries, allowing you to
provide a fallback value instead of getting an error. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.IfMissing}{documentation}
for details.
}
