% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{IfNot}
\alias{IfNot}
\title{IfNot query operation}
\usage{
IfNot(value = NULL, ...)
}
\arguments{
\item{value}{Optional value to use}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
A query operation providing a value to use for "false-ish" values in a vector (empty strings, zero numeric values,
or false Boolean values). See the
Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.IfNot}{documentation} for
details.
}
