% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{IfNot}
\alias{IfNot}
\title{IfNot query operation}
\usage{
IfNot(value = NULL, ...)
}
\arguments{
\item{value}{Optional value to use for replacement. If NULL, uses the default replacement value.}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
A query operation providing a value to use for "false-ish" values in a vector.
This replaces empty strings, zero numeric values, or false Boolean values with the specified value.
This is useful for handling default or missing values in data without completely replacing the property.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.IfNot}{documentation} for
details.
}
