% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquireInsertDataFrameToRedshift.R
\name{InsertDataFrameToRedshift}
\alias{InsertDataFrameToRedshift}
\title{Creates a Redshift table with the same contents as the given R data frame}
\usage{
InsertDataFrameToRedshift(df, tableName, columnTypes, schema = "development",
  clean = TRUE, dbID, s3ID, yamlConfig = "../db.yml")
}
\arguments{
\item{tableName}{The name of the table to create or append to}

\item{schema}{Character vector with length 1 with the name of the DB schema were the table will be created}

\item{clean}{Boolean indicating whether or not to remove existing records from the table before inserting}

\item{query}{Character vector with length 1. Can be either a SQL query or a path to a text file containing a SQL query}

\item{id}{The name of the yaml group containing the database credentials}

\item{yamlFile}{The path to the yaml file}
}
\value{
The function will return a boolean value indicating if the process completed successfully
}
\description{
Given an R data frame, or equivalently, any data.frame like object, the function will create a redshift table tthat mimics the contents of the R object.
This copy is done by first copying then object onto and s3 bucket as a .csv, and then uploading to redshift from there.

The yaml file must contain host, dbname, s3 bucket name, user and password.
}
\author{
João Monteiro and Henrique Cabral
}

