% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{install_daf_packages}
\alias{install_daf_packages}
\title{Install Julia packages needed for DataAxesFormats and TanayLabUtilities}
\usage{
install_daf_packages(force_dev = FALSE)
}
\arguments{
\item{force_dev}{(Default=FALSE) Whether to force dev versions of packages}
}
\value{
No return value, called for side effects.
}
\description{
Install Julia packages needed for DataAxesFormats and TanayLabUtilities
}
