% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{is_axis_query}
\alias{is_axis_query}
\title{Check if a query returns axis entries}
\usage{
is_axis_query(query)
}
\arguments{
\item{query}{Query string or query object}
}
\value{
TRUE if the query returns axis entries, FALSE otherwise
}
\description{
Determines whether a query will return axis entries (names) as opposed to
scalar values, vectors, or matrices.
}
\details{
This is useful for determining the expected result type of a query
before executing it. A query that returns axis entries can be used as a filter
for other queries.
}
