% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daf.R
\name{is_daf}
\alias{is_daf}
\title{Check if object is a Daf}
\usage{
is_daf(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
TRUE if x is a Daf object, FALSE otherwise
}
\description{
Check if object is a Daf
}
