% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{IsEqual}
\alias{IsEqual}
\title{IsEqual query operation}
\usage{
IsEqual(value, ...)
}
\arguments{
\item{value}{Value to compare against}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Equality is used for two purposes: As a comparison operator, similar to \code{IsLess} except that uses \code{=} instead of
\code{<} for the comparison; and To select a single entry from a vector. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAccessFormats.Queries.IsEqual}{documentation} for
details.
}
