% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{IsLess}
\alias{IsLess}
\title{IsLess query operation}
\usage{
IsLess(value, ...)
}
\arguments{
\item{value}{Value to compare against}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
A query operation for converting a vector value to a Boolean mask by comparing it some value. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAccessFormats.Queries.IsLess}{documentation} for details.
}
