% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{IsMatch}
\alias{IsMatch}
\title{IsMatch query operation}
\usage{
IsMatch(value, ...)
}
\arguments{
\item{value}{Regular expression pattern to match against}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Similar to \code{IsLess} except that the compared values must be strings, and the mask
is of the values that match the given regular expression. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.IsMatch}{documentation} for
details.
}
