% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{IsNotEqual}
\alias{IsNotEqual}
\title{IsNotEqual query operation}
\usage{
IsNotEqual(value, ...)
}
\arguments{
\item{value}{Value to compare against}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Similar to \code{IsLess} except that uses \code{!=} instead of \code{<} for the comparison. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.IsNotEqual}{documentation} for
details.
}
