% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{IsNotMatch}
\alias{IsNotMatch}
\title{IsNotMatch query operation}
\usage{
IsNotMatch(value, ...)
}
\arguments{
\item{value}{Regular expression pattern to not match against}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Similar to \code{IsMatch} except that looks for entries that do not match the pattern. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAccessFormats.Queries.IsNotMatch}{documentation} for
details.
}
