% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{matrices_set}
\alias{matrices_set}
\title{Names of matrices for an axis pair, sorted.}
\usage{
matrices_set(daf, rows_axis, cols_axis)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{rows_axis}{Row-axis name.}

\item{cols_axis}{Column-axis name.}
}
\value{
Character vector.
}
\description{
Names of matrices for an axis pair, sorted.
}
