% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{matrices_set}
\alias{matrices_set}
\title{Get set of matrix names for axes in a Daf object}
\usage{
matrices_set(daf, rows_axis, columns_axis, relayout = TRUE)
}
\arguments{
\item{daf}{A Daf object}

\item{rows_axis}{Name of rows axis}

\item{columns_axis}{Name of columns axis}

\item{relayout}{Whether to include matrices with flipped axes (TRUE by default)}
}
\value{
A character vector of matrix property names
}
\description{
Returns the names of all matrix properties for the specified pair of axes.
}
\details{
This function provides the complete set of available matrix properties
for specific axes that can be retrieved using \code{get_matrix()}.
If \code{relayout} is TRUE, matrices stored with the axes flipped are also included.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.matrices_set}{documentation} for details.
}
