% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{matrix_version_counter}
\alias{matrix_version_counter}
\title{Get matrix version counter}
\usage{
matrix_version_counter(daf, rows_axis, columns_axis, name)
}
\arguments{
\item{daf}{A Daf object}

\item{rows_axis}{Name of the rows axis}

\item{columns_axis}{Name of the columns axis}

\item{name}{Name of the matrix property}
}
\value{
An integer version counter
}
\description{
Returns the version counter for a matrix property, which is incremented when the matrix is modified.
This is useful for cache invalidation.
}
\details{
The version counter is incremented whenever the matrix data is modified.
This can be used to detect changes and invalidate caches.
}
