% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Max}
\alias{Max}
\title{Max query operation}
\usage{
Max(...)
}
\arguments{
\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Reduction operation that returns the maximal element. This operation reduces the
dimensionality of the data: a matrix becomes a vector (maximum of each column),
and a vector becomes a scalar (maximum of all elements). See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/operations.html#DataAxesFormats.Operations.Max}{documentation} for
details.
}
