% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Mean}
\alias{Mean}
\title{Mean query operation}
\usage{
Mean(...)
}
\arguments{
\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Reduction operation that returns the mean value. This operation reduces the
dimensionality of the data: a matrix becomes a vector (mean of each column),
and a vector becomes a scalar (mean of all elements). See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Operations.Mean}{documentation} for
details.
}
