% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Median}
\alias{Median}
\title{Median query operation}
\usage{
Median(type = NULL, ...)
}
\arguments{
\item{type}{Optional result type (e.g., "Float64"). If NULL, the default type is used.}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Reduction operation that returns the median value. This operation reduces the
dimensionality of the data: a matrix becomes a vector (median of each column),
and a vector becomes a scalar (median of all elements). See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/operations.html#DataAxesFormats.Operations.Median}{documentation} for
details.
}
