% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{memory_daf}
\alias{memory_daf}
\title{Create a Daf object with in-memory storage}
\usage{
memory_daf(name = "memory")
}
\arguments{
\item{name}{The name of the Daf object (default: "memory")}
}
\value{
A Daf object with in-memory storage
}
\description{
This function creates a Daf object that stores data in memory. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/memory_format.html}{documentation} for details.
}
