% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Min}
\alias{Min}
\title{Min query operation}
\usage{
Min(...)
}
\arguments{
\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Reduction operation that returns the minimal element. This operation reduces the
dimensionality of the data: a matrix becomes a vector (minimum of each column),
and a vector becomes a scalar (minimum of all elements). See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Operations.Min}{documentation} for
details.
}
