% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{mmap_real}
\alias{mmap_real}
\alias{mmap_int}
\alias{mmap_lgl}
\title{Open a binary file as a read-only mmap-backed numeric vector.}
\usage{
mmap_real(path, length)

mmap_int(path, length)

mmap_lgl(path, length)
}
\description{
Length is the number of \code{double} elements; the file must contain at least
\code{length * 8} bytes.
}
