% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Names}
\alias{Names}
\title{Names query operation}
\usage{
Names(kind = NULL, ...)
}
\arguments{
\item{kind}{Optional string specifying the kind}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
A query operation for looking up a set of names. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.Names}{documentation}
for details.
}
