% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Names}
\alias{Names}
\title{Names query operation}
\usage{
Names(kind = NULL, ...)
}
\arguments{
\item{kind}{Optional string specifying the kind of names to look up (e.g., "axes", "scalars")}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
A query operation for looking up a set of names in a Daf object.
This operation retrieves metadata names such as axis names, property names, or scalar names.
It is often used to discover what data is available in a Daf object.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.Names}{documentation}
for details.
}
