% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_store.R
\name{new_http_store}
\alias{new_http_store}
\title{Construct a read-only \link{ZarrStore} over HTTP.}
\usage{
new_http_store(url)
}
\arguments{
\item{url}{Root URL of a Zarr v2 directory served over HTTP(S).}
}
\value{
An \code{HttpStore}.
}
\description{
Construct a read-only \link{ZarrStore} over HTTP.
}
