% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{open_daf}
\alias{open_daf}
\title{Open a Daf repository based on path}
\usage{
open_daf(path, mode = "r", name = NULL)
}
\arguments{
\item{path}{Path to the Daf repository}

\item{mode}{Mode to open the storage ("r" for read-only, "r+" for read-write)}

\item{name}{Optional name for the Daf object}
}
\value{
A Daf object (either files_daf or h5df)
}
\description{
This function determines whether to open a files-based Daf or an HDF5-based Daf
based on the file path.
}
