% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_parse.R
\name{parse_query}
\alias{parse_query}
\title{Parse a query string into an AST (list of \code{qop} nodes).}
\usage{
parse_query(query_string)
}
\arguments{
\item{query_string}{A character scalar.}
}
\value{
A list of AST node records.
}
\description{
Parse a query string into an AST (list of \code{qop} nodes).
}
