% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{parse_query}
\alias{parse_query}
\title{Parse a query string into a query object}
\usage{
parse_query(query_string)
}
\arguments{
\item{query_string}{String containing the query}
}
\value{
A query object
}
\description{
Parse a query string into a query object
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html}{documentation} for details.
}
