% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Quantile}
\alias{Quantile}
\title{Quantile query operation}
\usage{
Quantile(p = 0.5, ...)
}
\arguments{
\item{p}{Quantile to compute (between 0 and 1). Default is 0.5 (median).}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Reduction operation that returns the quantile value, that is, a value such that
a certain fraction of the values is lower. This operation reduces the
dimensionality of the data: a matrix becomes a vector (quantile of each column),
and a vector becomes a scalar (quantile of all elements). See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/operations.html#DataAxesFormats.Operations.Quantile}{documentation} for
details.
}
