% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_axis_name}
\alias{query_axis_name}
\title{Return the axis name implied by a query, if any.}
\usage{
query_axis_name(query_string)
}
\arguments{
\item{query_string}{A query string.}
}
\value{
A single axis name, or \code{NA_character_} if the query references
0 or 2+ axes.
}
\description{
Return the axis name implied by a query, if any.
}
\examples{
query_axis_name("@ cell : donor")
query_axis_name(". organism")
}
