% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_axis_name}
\alias{query_axis_name}
\title{Get the axis name from a query}
\usage{
query_axis_name(query)
}
\arguments{
\item{query}{Query string or query object}
}
\value{
The name of the axis as a character string
}
\description{
Returns the name of the axis that a query operates on.
}
\details{
This is useful for understanding which axis a query will affect
or for building compound queries programmatically.
}
