% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_requires_relayout}
\alias{query_requires_relayout}
\title{Check if a query requires relayout}
\usage{
query_requires_relayout(daf, query)
}
\arguments{
\item{daf}{A Daf object}

\item{query}{Query string or query object}
}
\value{
TRUE if the query requires relayout, FALSE otherwise
}
\description{
Determines whether executing a query on a Daf object would require a matrix relayout
operation. This is useful for performance optimization, as relayout can be expensive.
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.query_requires_relayout}{documentation} for details.
}
