% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_result_dimensions}
\alias{query_result_dimensions}
\title{Return the dimensionality of a query's result.}
\usage{
query_result_dimensions(query_string)
}
\arguments{
\item{query_string}{A query string.}
}
\value{
0L (scalar), 1L (vector / axis entries), 2L (matrix), or
\code{NA_integer_} if the query is ill-formed.
}
\description{
Return the dimensionality of a query's result.
}
