% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_result_dimensions}
\alias{query_result_dimensions}
\title{Get the number of dimensions of a query result}
\usage{
query_result_dimensions(query)
}
\arguments{
\item{query}{Query string or object}
}
\value{
Number of dimensions (-1 - names, 0 - scalar, 1 - vector, 2 - matrix)
}
\description{
Determines the dimensionality of the result that would be produced by applying a query.
This is useful for understanding what kind of data structure to expect from a query before
executing it, which can help with writing code that handles different result types.
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html}{documentation} for details.
}
