% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{registered_reductions}
\alias{registered_reductions}
\title{List registered reduction operations.}
\usage{
registered_reductions()
}
\value{
Sorted character vector of registered reduction names.
}
\description{
Returns the names of all currently registered reduction operations.
}
\examples{
head(registered_reductions())
register_reduction("Median_example", function(x, ...) median(x), overwrite = TRUE)
"Median_example" \%in\% registered_reductions()
}
