% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{relayout_matrix}
\alias{relayout_matrix}
\title{Physically store the transposed layout of a matrix.}
\usage{
relayout_matrix(daf, rows_axis, cols_axis, name)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{rows_axis}{Row-axis name.}

\item{cols_axis}{Column-axis name.}

\item{name}{Matrix name.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
After this call, \code{get_matrix(cols_axis, rows_axis, name)} skips the
transpose-on-the-fly path.
}
