% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{relayout_matrix}
\alias{relayout_matrix}
\title{Relayout matrix in a Daf object}
\usage{
relayout_matrix(daf, rows_axis, columns_axis, name, overwrite = FALSE)
}
\arguments{
\item{daf}{A Daf object}

\item{rows_axis}{Name of rows axis}

\item{columns_axis}{Name of columns axis}

\item{name}{Name of the matrix property}

\item{overwrite}{Whether to overwrite if matrix already exists with flipped axes (FALSE by default)}
}
\value{
The Daf object (invisibly, for chaining operations)
}
\description{
Creates or updates a matrix property with flipped axes for more efficient access.
}
\details{
This function creates a transposed version of an existing matrix property,
allowing efficient access from either axis orientation.
If a matrix with the flipped axes already exists and \code{overwrite} is FALSE, an error will be raised.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/writers.html#DataAxesFormats.Writers.relayout_matrix!}{documentation} for details.
}
