% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{scalars_set}
\alias{scalars_set}
\title{Names of all scalars, sorted.}
\usage{
scalars_set(daf)
}
\arguments{
\item{daf}{A \code{DafReader}.}
}
\value{
Character vector.
}
\description{
Names of all scalars, sorted.
}
\examples{
# Mirrors readers.jl jldoctest at line 125.
scalars_set(example_cells_daf()) # "organism" "reference"
}
