% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scalars_set}
\alias{scalars_set}
\title{Get set of scalar names from a Daf object}
\usage{
scalars_set(daf)
}
\arguments{
\item{daf}{A Daf object}
}
\value{
A character vector of scalar property names
}
\description{
Returns the names of all scalar properties in the Daf data set.
}
\details{
This function provides the complete set of available scalar properties
that can be retrieved using \code{get_scalar()}.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.scalars_set}{documentation} for details.
}
