% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{set_matrix}
\alias{set_matrix}
\title{Set a matrix indexed by a pair of axes.}
\usage{
set_matrix(daf, rows_axis, cols_axis, name, mat, overwrite = FALSE)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{rows_axis}{Row-axis name.}

\item{cols_axis}{Column-axis name.}

\item{name}{Matrix name.}

\item{mat}{Dense \code{matrix}, or sparse \code{dgCMatrix} / \code{lgCMatrix}, of
shape \verb{axis_length(rows_axis) x axis_length(cols_axis)}.}

\item{overwrite}{See \code{set_scalar}.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
Set a matrix indexed by a pair of axes.
}
