% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{set_matrix}
\alias{set_matrix}
\title{Set a matrix indexed by a pair of axes.}
\usage{
set_matrix(daf, rows_axis, columns_axis, name, mat, overwrite = FALSE)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{rows_axis}{Row-axis name.}

\item{columns_axis}{Column-axis name.}

\item{name}{Matrix name.}

\item{mat}{Dense \code{matrix}, or sparse \code{dgCMatrix} / \code{lgCMatrix}, of
shape \verb{axis_length(rows_axis) x axis_length(columns_axis)}.}

\item{overwrite}{See \code{set_scalar}.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
Set a matrix indexed by a pair of axes.
}
\examples{
d <- memory_daf()
add_axis(d, "cell", c("c1", "c2"))
add_axis(d, "gene", c("g1", "g2", "g3"))
m <- matrix(1:6, nrow = 2, ncol = 3,
    dimnames = list(c("c1", "c2"), c("g1", "g2", "g3")))
set_matrix(d, "cell", "gene", "counts", m)
dim(get_matrix(d, "cell", "gene", "counts"))
}
