% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_matrix}
\alias{set_matrix}
\title{Set matrix in a Daf object}
\usage{
set_matrix(
  daf,
  rows_axis,
  columns_axis,
  name,
  value,
  overwrite = FALSE,
  relayout = TRUE
)
}
\arguments{
\item{daf}{A Daf object}

\item{rows_axis}{Name of rows axis}

\item{columns_axis}{Name of columns axis}

\item{name}{Name of the matrix property}

\item{value}{Matrix of values to set (cannot contain NA values)}

\item{overwrite}{Whether to overwrite if matrix already exists (FALSE by default)}

\item{relayout}{Whether to allow relayout with flipped axes (TRUE by default)}
}
\value{
The Daf object (invisibly, for chaining operations)
}
\description{
Sets a matrix property with the specified name for the given axes in the Daf data set.
}
\details{
This function creates or updates a matrix property in the Daf data set.
The dimensions of the matrix must match the lengths of the specified axes.
If the matrix already exists and \code{overwrite} is FALSE, an error will be raised.
If \code{relayout} is TRUE, the matrix will also be stored with axes flipped for faster access.
NA values are not supported in Daf.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/writers.html#DataAxesFormats.Writers.set_matrix!}{documentation} for details.
}
