% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_scalar}
\alias{set_scalar}
\title{Set scalar value in a Daf object}
\usage{
set_scalar(daf, name, value, overwrite = FALSE)
}
\arguments{
\item{daf}{A Daf object}

\item{name}{Name of the scalar property}

\item{value}{Value to set (cannot be NA)}

\item{overwrite}{Whether to overwrite if scalar already exists (FALSE by default)}
}
\value{
The Daf object (invisibly, for chaining operations)
}
\description{
Sets the value of a scalar property with the specified name in the Daf data set.
}
\details{
This function creates or updates a scalar property in the Daf data set.
If the scalar already exists and \code{overwrite} is FALSE, an error will be raised.
NA values are not supported in Daf.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/writers.html#DataAxesFormats.Writers.set_scalar!}{documentation} for details.
}
