% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{set_vector}
\alias{set_vector}
\title{Set a vector on an axis.}
\usage{
set_vector(daf, axis, name, vec, overwrite = FALSE)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{axis}{Axis name.}

\item{name}{Vector name.}

\item{vec}{Atomic vector of length \code{axis_length(daf, axis)}, or a
named vector whose names are a subset of the axis entries (reordered
into axis order at storage time).}

\item{overwrite}{See \code{set_scalar}.}
}
\value{
Invisibly the input \code{daf}.
}
\description{
Set a vector on an axis.
}
