% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_vector}
\alias{set_vector}
\title{Set vector in a Daf object}
\usage{
set_vector(daf, axis, name, value, overwrite = FALSE)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Axis name}

\item{name}{Name of the vector property}

\item{value}{Vector of values to set (cannot contain NA values)}

\item{overwrite}{Whether to overwrite if vector already exists (FALSE by default)}
}
\value{
The Daf object (invisibly, for chaining operations)
}
\description{
Sets a vector property with the specified name for an axis in the Daf data set.
}
\details{
This function creates or updates a vector property in the Daf data set.
The length of the vector must match the length of the axis.
If the vector already exists and \code{overwrite} is FALSE, an error will be raised.
NA values are not supported in Daf.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/writers.html#DataAxesFormats.Writers.set_vector!}{documentation} for details.
}
