% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{StdN}
\alias{StdN}
\title{StdN query operation}
\usage{
StdN(type = NULL, eps = NULL, ...)
}
\arguments{
\item{type}{Optional result type (e.g., "Float64"). If NULL, the default type is used.}

\item{eps}{Optional small value added to the mean before dividing. If NULL, the default is used.}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Reduction operation that returns the standard deviation of the values, normalized
(divided) by the mean of the values. This operation reduces the dimensionality
of the data: a matrix becomes a vector (normalized standard deviation of each column),
and a vector becomes a scalar (normalized standard deviation of all elements). See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/operations.html#DataAxesFormats.Operations.Std}{documentation} for details.
}
