% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{Sum}
\alias{Sum}
\title{Sum query operation}
\usage{
Sum(...)
}
\arguments{
\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Reduction operation that sums elements. This operation reduces the
dimensionality of the data: a matrix becomes a vector (sum of each column),
and a vector becomes a scalar (sum of all elements). See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Operations.Sum}{documentation} for
details.
}
