% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{unescape_value}
\alias{unescape_value}
\title{Unescape a value from a query string}
\usage{
unescape_value(value)
}
\arguments{
\item{value}{A character string value to unescape}
}
\value{
The unescaped value as a character string
}
\description{
Reverses the escaping done by \code{escape_value}, restoring the original value.
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Queries.unescape_value}{documentation} for details.
}
