% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{VarN}
\alias{VarN}
\title{VarN query operation}
\usage{
VarN(...)
}
\arguments{
\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
Reduction operation that returns the variance of the values, normalized (divided)
by the mean of the values. This operation reduces the dimensionality of the data:
a matrix becomes a vector (normalized variance of each column), and a vector becomes
a scalar (normalized variance of all elements). See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/queries.html#DataAxesFormats.Operations.Var}{documentation}
for details.
}
