% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{vector_version_counter}
\alias{vector_version_counter}
\title{Get vector version counter}
\usage{
vector_version_counter(daf, axis, name)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis}

\item{name}{Name of the vector property}
}
\value{
An integer version counter
}
\description{
Returns the version counter for a vector property, which is incremented when the vector is modified.
This is useful for cache invalidation.
}
\details{
The version counter is incremented whenever the vector data is modified.
This can be used to detect changes and invalidate caches.
}
