% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{vectors_set}
\alias{vectors_set}
\title{Names of vectors on an axis, sorted.}
\usage{
vectors_set(daf, axis)
}
\arguments{
\item{daf}{A \code{DafReader}.}

\item{axis}{Axis name.}
}
\value{
Character vector.
}
\description{
Names of vectors on an axis, sorted.
}
\examples{
# Mirrors readers.jl jldoctest at line 595.
vectors_set(example_cells_daf(), "cell") # "donor" "experiment"
}
