% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{vectors_set}
\alias{vectors_set}
\title{Get set of vector names for an axis in a Daf object}
\usage{
vectors_set(daf, axis)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis}
}
\value{
A character vector of vector property names
}
\description{
Returns the names of all vector properties for the specified axis.
}
\details{
This function provides the complete set of available vector properties
for a specific axis that can be retrieved using \code{get_vector()}.
Vector properties store one-dimensional data along a specific axis.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/readers.html#DataAxesFormats.Readers.vectors_set}{documentation} for details.
}
