% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\docType{data}
\name{VIEW_ALL_AXES}
\alias{VIEW_ALL_AXES}
\title{View all axes specifier}
\format{
A named list representing the pair ALL_AXES => "="
}
\usage{
VIEW_ALL_AXES
}
\description{
A pair to use in the \code{axes} parameter of \code{viewer} to expose all base axes with their original names.
See the Julia documentation for details.
}
\references{
\url{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/views.html#DataAxesFormats.Views.VIEW_ALL_AXES}
}
\keyword{datasets}
