% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\docType{data}
\name{VIEW_ALL_DATA}
\alias{VIEW_ALL_DATA}
\title{Default view spec: expose all data (scalars + vectors + matrices) as-is.}
\format{
An object of class \code{list} of length 3.
}
\usage{
VIEW_ALL_DATA
}
\description{
A convenience list combining \link{VIEW_ALL_SCALARS}, \link{VIEW_ALL_VECTORS}, and
\link{VIEW_ALL_MATRICES}. Pass as the \code{data} argument to \code{viewer()} to expose
every data item from the base daf unchanged.
}
\seealso{
\code{\link[=viewer]{viewer()}}, \link{VIEW_ALL_SCALARS}, \link{VIEW_ALL_VECTORS},
\link{VIEW_ALL_MATRICES}
}
\keyword{datasets}
