% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\docType{data}
\name{VIEW_ALL_DATA}
\alias{VIEW_ALL_DATA}
\title{View all data specifier}
\format{
A list containing VIEW_ALL_SCALARS, VIEW_ALL_VECTORS, and VIEW_ALL_MATRICES
}
\usage{
VIEW_ALL_DATA
}
\description{
A list of pairs to use in the \code{data} parameter of \code{viewer} to expose all base data
(scalars, vectors, and matrices) with their original names.
See the Julia documentation for details.
}
\references{
\url{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/views.html#DataAxesFormats.Views.VIEW_ALL_DATA}
}
\keyword{datasets}
