% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\docType{data}
\name{VIEW_ALL_MATRICES}
\alias{VIEW_ALL_MATRICES}
\title{View all matrices specifier}
\format{
A named list representing the pair ALL_MATRICES => "="
}
\usage{
VIEW_ALL_MATRICES
}
\description{
A pair to use in the \code{data} parameter of \code{viewer} to expose all base matrices with their original names.
See the Julia documentation for details.
}
\references{
\url{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/views.html#DataAxesFormats.Views.VIEW_ALL_MATRICES}
}
\keyword{datasets}
