% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_daf.R
\docType{data}
\name{VIEW_ALL_SCALARS}
\alias{VIEW_ALL_SCALARS}
\title{Default view spec: expose all scalars as-is.}
\format{
An object of class \code{list} of length 1.
}
\usage{
VIEW_ALL_SCALARS
}
\description{
A pre-built \code{data} item for \code{viewer()} that exposes every scalar of the base
daf unchanged. Equivalent to \code{list(list(ALL_SCALARS, "="))}.
}
\examples{
d <- memory_daf()
set_scalar(d, "organism", "human")
v <- viewer(d, data = list(VIEW_ALL_SCALARS))
scalars_set(v)
}
\seealso{
\code{\link[=viewer]{viewer()}}, \link{ALL_SCALARS}
}
\keyword{datasets}
