% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{viewer}
\alias{viewer}
\title{Create a read-only view of a Daf data set}
\usage{
viewer(daf, name = NULL, axes = NULL, data = NULL)
}
\arguments{
\item{daf}{A Daf object}

\item{name}{Optional name for the view}

\item{axes}{Named list specifying axes to expose}

\item{data}{Named list specifying data to expose}
}
\value{
A read-only Daf object
}
\description{
Wrap a Daf data set with a read-only DafView
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.1.2/views.html#DataAxesFormats.Views.viewer}{documentation} for details.
}
