% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr_store.R
\name{ZarrStore}
\alias{ZarrStore}
\alias{DirStore}
\alias{DictStore}
\alias{MmapZipStore}
\title{Zarr v2 Store abstract base class.}
\usage{
ZarrStore()

DirStore(root = character(0))

DictStore(env = NULL)

MmapZipStore(path = character(0))
}
\description{
Uniform read/write API for byte content keyed by path strings.
Concrete implementations: \link{DirStore}, \link{DictStore}, \link{MmapZipStore}.
}
